==> src/old_scripts/2. Try_data4.Rmd <==
---
title: "TRY database"
author: "Jelyn Gerkema"
date: "2023-11-17"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## 0. Set up
```{r set up}
source("src/settings.R")
source("src/functions.R")

try_categorical <- read_excel("data/raw/TRY/TRY_Categorical_Traits_Lookup_Table_2012_03_17_TestRelease.xlsx")
try_taxonomic <- read_excel("data/raw/TRY/try_taxonomic.xlsx")
library(WorldFlora)

classification <- read_delim("data/raw/classification.csv", 
    delim = "\t", escape_double = FALSE, 
    trim_ws = TRUE)

library(rtry)

```


```{r try}


path_to_data <- "data/raw/TRY/29649.txt"

TRYdata1 <- rtry_import(path_to_data)

#tmp_unfiltered <- rtry_select_row(TRYdata1, DataID %in% 413)

#tmp_unfiltered <- rtry_explore(tmp_unfiltered, DataID, DataName, OriglName, OrigValueStr, OrigUnitStr, StdValue, Comment, sortBy = OrigValueStr)


try_ls <- rtry_exclude(TRYdata1, (DataID %in% 413) & 
                         (OrigValueStr %in% c("13 weeks after sowing, 8 weeks after start of treatments",
          "check", "compound", "immature", "intermed", "junvenil", "juvenil", "Juvenil", "juvenil (3 years)",
          "juvenil: 0.5-2.5 m high", "juvenile", "Juvenile", "juvenile, 11-14 weeks", "Juvenile, 15 years",
          "juvenile, 6 weeks", "S", "sapling", "Sapling (1 - 10 y)", "saplings", "saplings (height 1-2m)",
          "saplings, 5 - 10 years", "second years after sowing seeds (relatively juvenile for most of species)",
          "seedling", "Seedling", "Seedling (0 - 1 y)", "seedlings", "seedlings, < 1/2 year", 
          "seedlings, 1st year", "seedlings, 6 month old", "T", "T/F", "T?", "Y", "young", "Young leaves", 
          "young tree", "young trees at  0.5-3m height", "young, not seedlings")), baseOn = ObservationID)


#tmp_unfiltered <- rtry_select_row(TRYdata1, DataID %in% 327)

#tmp_unfiltered <- rtry_explore(tmp_unfiltered, DataID, DataName, OriglName, OrigValueStr, OrigUnitStr, StdValue, Comment, sortBy = OrigValueStr)

try_ls_exp <- rtry_exclude(try_ls, (DataID %in% 327) & 
                         (OrigValueStr %in% c("BAU-Abandonnée/Nardaie Reposoir", 
          "BAU-Abandonnée/Nardaie non paturée", "BAU-Abandonnée/prairie grasse non paturée", 
          "BAU-Abandonée/Seslérie non paturée.", "BAU-Nardaie paturée.", 
          "BAU-Paturée Intermediaire Grasse-seslérie",
          "BAU-Prairie grasse paturée", "BAU-Prairie à Seslérie paturée", "Branch Bag", "C", "CG", "Chamber",
          "Climate Chamber", "Climate chamber, non-limiting conditions, (cf. dataset reference)", 
          "Common Garden", "Common Garden Experiment", "Common garden", "Common garden trees", 
          "Controlled climate chamber", "Dry Plot", "Experimental, Glasshouse", "FACE",
          "FE", "Field Experiment", "Fully open overstory 90 days, seedling", 
          "Fully open overstory, seedling", "G", "GH", "Glass house", "Glasshouse", 
          "Glasshouse experiment", "Greehouse", "Green house", "Greenhouse", "Greenhouse plants",
          "Greenhouse, Indiana University", "Greenhouse, grrowth container", 
          "Greenhouse: highlight_highpH_competition", "Greenhouse: highlight_highpH_nocompetition",
          "Greenhouse: highlight_lowpH_competition", "Greenhouse: highlight_lowpH_nocompetition",
          "Greenhouse: lowleight_lowpH_competition", "Greenhouse: lowlight_highpH_competition",
          "Greenhouse: lowlight_highpH_nocompetition", "Greenhouse: lowlight_lowpH_nocompetition",
          "Grown from seed sown in outdoor field experiment", "Growth Chamber", "Growth chamber", 
          "Growth exp", "Harte warming experiment (montane meadow)", 
          "I do not like this paper. Results looks questionable", 
          "Irrigation and N fertilisation (100 kg/ha)", "LAU_Grazed/Abandoned", 
          "LAU_Never ploughed/grazed", "LAU_Never ploughed/grazed highland", 
          "LAU_Never ploughed/grazed slopes", "LAU_Never ploughed/mown", "LAU_Ploughed/grazed",
          "LAU_Ploughed/mown", "LAU_Ploughed/mown and fertilized", 
          "Large potted monoculture mesocosms, outside, soil from local meadow,  grazing and harvests emulated",
          "Mature cultivated", "Monoculture", "Natural/C", "OTC", "OTC O3", "OTC field", 
          "Open Top", "Open top chambers", "Other", "Outdoor cultivation in 30x30cm pots, standardized substrate",
          "Outdoor cultivations (4 repititions) grown in 30x30cm pots with standardized substrate. All pot ...",
          "PM", "PU", "Partially open overstory 90 days, seedling", "Planted trees", "Planted vines",
          "Pot exp", "Pot-grown", "Pots outside", "Siemianice", "Undisturbed, seedling", 
          "University of California Botanical Garden", "VER_permanent meadow mown and fertilized",
          "ER_permanent meadows mown and fertilized", "WTC", "botanical garden environment",
          "branch bag", "climate chamber", "climate chambers", "comm.garden.Spain", 
          "common garden in growth containers with soil corresponding to seed origin", 
          "common garden, understorey", "controlled environment", "controlled environment room", 
          "cultivated", "disturbed soil treatment; coniferous forest", 
          "disturbed soil treatment; deciduous alluvial forest", "disturbed soil treatment; fallow wet meadow",
          "drought treatment", "dry season measurements (see also #659)", "experimental",
          "experimental treatment", 
          "experimental warming with open top chambers during the snow-free period, increasing daily mean  ...",
          "experimental, in pots outside", "experimental, outside in pots", "field experiment",
          "forest fertilization", "glasshouse, climate controlled", 
          "glasshouse, plastic tubes 10cm diameter, 100 cm depth", "grassland common garden", "greenhouse",
          "greenhouse - Northern Arizona University (standard controlled environment)", "growth chamber",
          "growth-chamber", "growth_chamber", "hydroponic", "lab", "mesocosm", "mesocosm, 50l pots",
          "mini-ecosystem", "natural environment, high warming +4C, preccipitation ambient",
          "natural environment, high warming +4C, preccipitation ambient +50%", 
          "natural environment, high warming +4C, preccipitation ambient -50%",
          "natural environment, low warming +1.5C, preccipitation ambient",
          "natural environment, low warming +1.5C, preccipitation ambient +50%", 
          "natural environment, low warming +1.5C, preccipitation ambient -50%", 
          "natural environment, medium warming +2.5C, preccipitation ambient",
          "natural environment, medium warming +2.5C, preccipitation ambient +50%",
          "natural environment, medium warming +2.5C, preccipitation ambient -50%",
          "natural environment, no warming, preccipitation ambient +50%",
          "natural environment, no warming, preccipitation ambient -50%",
          "natural grassland, experimental nutrient NP addition", "nursery",
          "nutrient addition experiment", "open-sided growth chamber", "open-top chamber",
          "open-top chambers", "pot", "pots, outside in natural environment", "shade houses",
          "smelter polluted environment", "water stress experiment", "water treatment", 
          "whole-tree chamber", "plantation")), baseOn = ObservationID)

#tmp_unfiltered <- rtry_select_row(TRYdata1, DataID %in% 1961)

#tmp_unfiltered <- rtry_explore(tmp_unfiltered, DataID, DataName, OriglName, OrigValueStr, OrigUnitStr, StdValue, Comment, sortBy = OrigValueStr)

try_ls_exp_hlth <- rtry_exclude(try_ls_exp, (DataID %in% 1961) & 
                         (OrigValueStr %in% c("Dead")), baseOn = ObservationID)

#try_covariate <- try_filtered %>%
 # filter(is.na(TraitID))

#covariates <- unique(try_covariate$DataName) %>% as.data.frame()

try_filtered <- rtry_remove_dup(try_ls_exp_hlth)

#tmp_unfiltered <- rtry_explore(try_filtered, DataID, DataName, TraitID, TraitName, ErrorRisk, sortBy = ErrorRisk)

try_filtered <- rtry_exclude(try_filtered, ErrorRisk >= 3, baseOn = ObsDataID) 

try_clean <- try_filtered %>%
  mutate(OrigValueStr =  case_when(OrigValueStr == "c3" ~ "C3",
                             OrigValueStr == "C3?" ~ "C3",
                             OrigValueStr == "c4" ~ "C4",
                             OrigValueStr == "C4?" ~ "C4",
                             TRUE ~ OrigValueStr)) %>%
  filter(DataID %notin% c(8178, 9780, 1629, 1739, 974, 2526, 2760, 2527, 2761, 
                          9906, 2762, 9578, 9577, 9576, 9575, 9574, 3710, 2958, 
                          9447, 8651, 8643, 8648, 8642, 8650, 8647, 8646, 8649, 
                          8645, 8644, 2373, 2380, 2365, 2367, 2375, 2381, 3809, 
                          3812, 3814, 3811, 3816, 225, 2082, 2142, 3990, 3991, 
                          2412, 2413, 2414, 904, 1271, 9926, 9931, 27, 9898, 
                          9899, 9900, 9901, 476, 2506, 829, 2499, 2500, 1027, 
                          1028, 2311, 4142, 3974, 3979, 3980, 3978, 3976, 3975, 
                          3977)) %>%
  filter(TraitID %in% c(3, 4, 13, 14, 15, 18, 22, 24, 26, 27, 46, 47, 48, 53, 54, 
                        59, 70, 71, 83, 89, 145, 146, 164, 246, 334, 785, 819, 839, 
                        1047, 1080, 1181, 1256, 3106, 3113, 3117, 3588, 3801)) %>%
  drop_na(StdValue) %>%
  select(AccSpeciesName, TraitID, ObservationID, TraitName, StdValue,
         Reference, LastName, FirstName) %>%
  group_by(AccSpeciesName, ObservationID, TraitID, TraitName,
         Reference, LastName, FirstName) %>%
  summarise(across(where(is.numeric), ~mean(., na.rm = TRUE))) %>%
  ungroup()

# De opgeschoonde taxonomic dataset van TRY -- vanaf hier tot poging 2 wellicht verwijderen
try_taxonomic_2 <- try_taxonomic %>%
  rename(AccSpeciesID = TRY_AccSpeciesID,
         AccSpeciesName = TRY_AccSpeciesName) %>%
  select(AccSpeciesID, AccSpeciesName, Genus, Family) %>%
  distinct()


# Voor de BHPMF wil ik een dataset met unieke soortsnamen en informatie over 
# het genus en het geslacht. De Try database heeft een eigen taxonomische dataset.
# Daar staan echter dubbele soortsnamen in. 

# Hier wordt de gecleande try database gematcht met de try taxonomic dataset,
# Soorten zonder genus of family informatie vallen af (n = 1864)
try_tax_clean <- try_clean %>%
  select(AccSpeciesName) %>%
  distinct() %>%
  left_join(try_taxonomic_2) %>%
  drop_na(Genus) %>%
  drop_na(Family) %>%
  filter(Family %notin% c("Unknown", "unknown")) 

save(try_clean, try_filtered,
     file = "data/processed/try_save4112.RData")

############################# Poging 2
try_clean2 <- try_clean %>%
  filter(str_count(AccSpeciesName, "\\S+") > 1) %>%
  mutate(AccSpeciesName = tolower(AccSpeciesName)) %>%
  mutate(AccSpeciesName = paste0(tools::toTitleCase(word(AccSpeciesName, 1)), " ", word(AccSpeciesName, 2, -1))) 



try_wf_clean2 <- classification %>%
  mutate(scientificName = str_replace_all(scientificName, "  ", " ")) %>%
  filter(scientificName %in% c(try_clean2$AccSpeciesName))

not_included_in_wf <- try_clean2 %>%
  filter(AccSpeciesName %notin% c(try_wf_clean2$scientificName)) %>%
  select(AccSpeciesName) %>%
  distinct() %>%
  separate(AccSpeciesName, into = c("genus", "sp"), extra = "merge", remove = F) %>%
  filter(sp %notin% c("sp", "sp.", "abandoned area", "african shrubs", "alba preta",
                      "alluvial", "alluvial meadow", "araba tuua", "arc pp144",
                      "arc pp146", "average", "broad-leaved", "broadleaf", 
                      "catinga species", "cf.", "cff", "chocolate","cottage garden hybrid'",
                      "crop", "cultivar", "cultivars", "da", "da-mata", "da mata",
                      "da rosca", "da terra firme", "dark green", "darkcrasa", "de",
                      "deciduous forest", "deciduous trees", "evergreen trees",
                      "field annual", "findlay's blue'", "fisch.", "fitzgerald river",
                      "forest species", "form taxon 'filifolia'", "fuzzy", 
                      "gorky", "grande", "grass", "grasses", "grassland", 
                      "herbaceous", "hutch.", "hybrid", "hybrida_x", "hybridum",
                      "hybridum_x", "hybridus_x", "l.", "laeve", "leaf", "leafea",
                      "limestone", "lookalike", "m.", "meadow", "meadow steppe",
                      "melkboom", "mill.", "mosaic", "numm.", "oak", "ob", "palm",
                      "para", "pasture", "preta", "preto", "preto folha grauda",
                      "preto folha miuda", "Rosette forb", "shrub", "shrubs and trees",
                      "sj", "skinny", "skinny2", "skinnyret", "small", "ss", "steppe",
                      "tallmorph", "terete-leaved", "terra", "tree", "tree co2", 
                      "tree field", "trees", "wedding day'"
                      )) %>%
  filter(genus %notin% c("Mountain", "may", "Lauraceae", "Southafrica", "Tropical",
                         "")) %>%
  filter(AccSpeciesName %notin% c("Paeonia 'anne rosse'", "Mid liana", 
                                  "Mata mata", "Mata mata preto", 
                                  "Mata mata si", "Buxuxu new", "Buxuxu old",
                                  "Sold pus", "Pubescent rosette", "Carex saw",
                                  "Eri sch", "Ucuuba t folha", "Death valley annual", 
                                  "Death valley annuals", "Deciduous chaparral shrubs"
                                  )) # mata mata is a turtle species
# dit zijn alle soorten zonder directe match.
# Daar zit troep bij. Dat wil ik eruit hebben. 

to_be_matched <- not_included_in_wf %>%
  pull(AccSpeciesName)

matched_with_wf <- WFO.match(spec.data = to_be_matched, WFO.data = classification)

matched_clean <- matched_with_wf %>%
  filter(Matched == TRUE) %>%
  filter(taxonRank == "species") %>%
  relocate(scientificName, Fuzzy.dist) %>%
  filter(Subseq == 1) 

matched_clean2 <- matched_clean %>%
  select(c(taxonID, scientificName, spec.name)) %>%
  rename(new_name = scientificName, AccSpeciesName = spec.name,
         acceptedNameUsageID = taxonID) %>%
   mutate(acceptedNameUsageID = recode(acceptedNameUsageID, "wfo-0000667939" = "wfo-0001067485"))



# Oké wat wil ik. Een gedeelte van de try database heeft een directe match 
# met de classification dataset. Dus daarvan heb ik de soor

try_wf_tax_clean <- try_wf_clean2 

### AAAAAAAAAAAAAAh het lukt vandaag niet. Mijn hoofd is wazig, warrig, snel 
# afgeleid. Ik weet het even niet. Ik heb te veel energie in mijn lijf, te veel
# onrust. Een gejaagd gevoel. Heb ik teveel suiker gegeten, te weinig water gedronken
# ik weet het niet. ik wil dit zo graag afmaken, ik wil zo graag een stap verder zijn.
# maar mijn hoofd loopt even vast. Alsof hij te vol zit. Ik wil afleiding, maar de tijd
# kruipt en vliegt. Ik wil in de zone zitten maar het lukt niet eens om te beginnen. 

# Er zitten dubbele entrees in. Maar maakt dat uit. Ik weet het niet. 


accepted_species <- try_wf_clean2 %>%
  filter(taxonomicStatus != "Synonym") %>%
  relocate(taxonRank, acceptedNameUsageID, taxonomicStatus) %>%
  select(scientificName, taxonID) %>%
  rename(acceptedNameUsageID = taxonID, AccSpeciesName = scientificName) %>%
  mutate(new_name = AccSpeciesName) %>%
  arrange(acceptedNameUsageID) %>% # Some species have multiple ID numbers, due
  # to a different associated author with the exact same name. With the 
  # last lines of codes, the duplicates are removed. 
  distinct(AccSpeciesName, .keep_all = TRUE)

length(unique(synonyms$scientificName))

synonyms <- try_wf_clean2 %>% # & !any(nomenclaturalStatus == "Illegitimate")
  filter(taxonomicStatus == "Synonym") %>%
  filter(scientificName %notin% c(accepted_species$AccSpeciesName)) %>% 
  # Filter out all the synonyms that have an actual species match. I am going to
  # assume that if an author writes down that name, it means the original, accepted 
  # species name and not a synonym of another species. 
  #filter(nomenclaturalStatus %notin% c( "Invalid", "Rejected")) %>%
  select(c(scientificName, acceptedNameUsageID, nomenclaturalStatus)) 

# First, the synonyms are extracted which have a unique name. To do so, first 
# the duplicate copies are removed (i.e. the speciesnames that occur more than once
# but still with the same acceptedNameUsageID)
single_synonyms <- synonyms %>%
  select(c(scientificName, acceptedNameUsageID)) %>%
  distinct() %>%
  group_by(scientificName) %>%
  filter(n() == 1) %>%
  ungroup()

multiple_synonyms <- synonyms %>%
  filter(scientificName %notin% c(single_synonyms$scientificName)) %>%
  #group_by(scientificName) %>%
  filter(nomenclaturalStatus %in% c(NA, "Valid")) 

single_synonyms_2 <- multiple_synonyms %>%
  group_by(scientificName) %>%
  filter(n() == 1)

single_synonyms_3 <- multiple_synonyms %>%
  filter(scientificName %notin% c(single_synonyms_2$scientificName)) %>%
  group_by(scientificName) %>%
  filter(any(!is.na(nomenclaturalStatus))) %>%
  drop_na() %>%
  filter(n() == 1)

all_synonyms <- bind_rows(single_synonyms, single_synonyms_2, single_synonyms_3) %>%
  mutate(include = "yes")


codes <- classification %>%
  select(taxonID, scientificName) %>%
  rename(acceptedNameUsageID = taxonID, new_name = scientificName) 

matched_synonyms <- full_join(all_synonyms, codes) %>%
  filter(include == "yes") %>%
  rename(AccSpeciesName = scientificName) 

matched_synonyms_filtered <- matched_synonyms %>%
  merge(accepted_species, by = c("new_name" ="new_name")) %>%
  filter(!acceptedNameUsageID.x != acceptedNameUsageID.y) %>%
  select(new_name, acceptedNameUsageID.x, AccSpeciesName.x) %>%
  rename(acceptedNameUsageID = acceptedNameUsageID.x,
         AccSpeciesName = AccSpeciesName.x)
  
  

combined <- bind_rows(matched_synonyms_filtered, accepted_species)

length(unique(combined$AccSpeciesName))


### IMPORTANT!! this is the dataset with the old species names, the wfo number, 
# and their new name. There is one thing that still needs fixing: 
# some species are present double. That is because in some cases, a species is 
# a species on its own but also a synonym with another. For simplicity sakes,
# I will assume that if a species is present in the TRY database, it is an accepted name
# and not a synonym. 
# In most cases, the old and new name are the same. But there are also some 
try_direct_match2 <- try_clean2 %>%
  filter(AccSpeciesName %in% c(try_wf_clean2$scientificName)) %>%
  select(AccSpeciesName) %>%
  distinct() %>%
  full_join(combined) %>% #92 synonym species had multiple matches and where removed
  #select(-c(nomenclaturalStatus)) %>%
  drop_na(new_name) #

matched_clean3 <- matched_clean2 %>%
  merge(accepted_species, by = c("new_name" ="new_name")) %>%
  filter(!acceptedNameUsageID.x != acceptedNameUsageID.y) %>%
  select(new_name, acceptedNameUsageID.x, AccSpeciesName.x) %>%
  rename(acceptedNameUsageID = acceptedNameUsageID.x,
         AccSpeciesName = AccSpeciesName.x)
  
matched_clean_final <- matched_clean2 %>%
  filter(new_name %notin% c(accepted_species$new_name)) %>%
  bind_rows(matched_clean3)

eindelijk_klaar <- bind_rows(try_direct_match2, matched_clean_final)

length(unique(eindelijk_klaar$new_name))


  
final_taxonomic_information <- classification %>%
  filter(taxonID %in% c(eindelijk_klaar$acceptedNameUsageID)) %>%
  select(taxonID, genus, family) %>%
  rename(acceptedNameUsageID = taxonID)

final_pls_be_final <- full_join(eindelijk_klaar, final_taxonomic_information) %>%
  drop_na()

final_species_list <- final_pls_be_final %>%
  select(new_name, genus) %>%
  distinct() #61334

hierarchy <- lookup_table(final_species_list$new_name, genus_column = "genus", by_species=F, 
                          family.tax = "apweb", missing_action = "NA") %>%

 select(family, order, group)  %>%
  distinct() 

 final_test <- full_join(final_pls_be_final, hierarchy) %>%
   mutate(order =  case_when(family == "Alzateaceae" ~ "Myrtales",
                            family == "Agdestidaceae" ~ "Caryophyllales",
                            family == "Bataceae"  ~ "Brassicales",
                            family == "Crypteroniaceae" ~ "Myrtales",
                            family == "Didymochlaenaceae" ~ "Polypodiales",
                            family ==  "Francoaceae" ~ "Geraniales",
                            family == "Griseliniaceae" ~ "Apiales",
                            family == "Isoetaceae" ~ "Isoetales",
                            family == "Hemidictyaceae" ~ "Polypodiales",
                            family == "Lonchitidaceae" ~ "Polypodiales",
                            family == "Macarthuriaceae" ~ "Caryophyllales",
                            family == "Mazaceae" ~ "Lamiales",
                            family == "Namaceae" ~ "Boraginales",
                            family == "Nyssaceae" ~ "Cornales",
                            family == "Ripogonaceae" ~ "Liliales",
                            family == "Petiveriaceae" ~ "Caryophyllales",
                            family == "Thomandersiaceae" ~ "Lamiales",
                            family == "Viburnaceae" ~ "Dipsacales",
                                  TRUE ~ order)) %>%
    filter(family %notin% c("Bryaceae", "Daltoniaceae", "Leucobryaceae",
                          "Pallaviciniaceae", "Pylaisiaceae", 
                          "Gymnomitriaceae", "Hypnaceae",
                          "Jungermanniaceae")) %>%
   drop_na(new_name) %>%
   select(-group)
 
length(unique(final_test$new_name)) 
 


#########@@@@@@@@@@@################## Einde poging 2 

test_try_tax2 <- try_taxonomic %>%
  filter(TRY_AccSpeciesName %in% c(not_included2$AccSpeciesName)) %>%
  select(TRY_AccSpeciesName, MatchedName) %>%
  distinct() %>%
  filter(str_count(MatchedName, "\\S+") > 1)

length(unique(test_try_tax2$TRY_AccSpeciesName))

temp6 <- try_taxonomic %>%
  filter(TRY_AccSpeciesName == "Adesmia volkmannii")

count <- test_try_tax2 %>%
  count(TRY_AccSpeciesName) %>%
  filter(n > 1)

temp <- test_try_tax2 %>%
  filter(TRY_AccSpeciesName %in% c(count$TRY_AccSpeciesName))

test$scientificName <- classification %>%
  filter(taxonID %in% c("wfo-0000511233")) 

length(unique(try_tax_clean2$scientificName)) #61551

length(unique(try_clean$AccSpeciesName)) #69324

try_tax_species_list <- try_tax_clean %>%
  select(AccSpeciesName) %>%
  distinct() %>%
  mutate(try_species = "Yes") # De lijst heeft een lengte van 67460
# De try_tax_clean heeft een lengte van 67952, er zitten dus soorten dubbel in.

# Om de soorten die er dubbel in staan eruit te halen, wordt de soortenlijst
# gematcht met de WFO taxonomische informatie dataframe. 
filter <- classification %>%
    filter(taxonomicStatus != "Synonym") %>%
  filter(taxonID %notin% c("wfo-1200010625", "wfo-1200010660", "wfo-0001233171")) %>% 
  # Remove entries of three species with the wrong family name 
  select(scientificName, family, genus) %>%
  rename(AccSpeciesName = scientificName) %>%
  left_join(try_tax_species_list) %>%
  distinct() %>%
  filter(try_species == "Yes") %>% #57249
  drop_na(family) # Er zitten 16 soorten in die geen family informatie hebben
  # Die worden op deze manier eruit gehaald


# Dit zijn de soorten die geen match hadden in de WF dataset. Als het enkel
# unieke namen waren geweest dan was het klaar geweest en hadden ze samengevoegd
# kunnen worden. Maar dat is niet het geval. Dus de laatste duplicaten moeten er
# nog uitgehaald worden. Dat zijn de soorten die er met meerdere familie namen 
# in staan. 
not_included <- try_tax_clean %>%
  filter(AccSpeciesName %notin% c(filter$AccSpeciesName)) 

length(unique(not_included$AccSpeciesName))

# Het aantal voorkomen van elke species naam wordt geteld, daarmee wordt de lijst
# met duplicaten gecreeërd. 
last_duplicates <- try_tax_clean %>%
  filter(AccSpeciesName %in% c(not_included$AccSpeciesName)) %>%
  count(AccSpeciesName) %>%
  filter(n > 1)

# De unieke soorten worden er ook uitgehaald
unique_species <- try_tax_clean %>%
  filter(AccSpeciesName %in% c(not_included$AccSpeciesName)) %>%
  filter(AccSpeciesName %notin% c(last_duplicates$AccSpeciesName)) %>%
  rename(family = Family, genus = Genus)

# Met de last_duplicates namen wordt de oorspronkelijke dataset gefiltered.
selected_duplicates <- try_tax_clean %>%
  filter(AccSpeciesName %in% c(last_duplicates$AccSpeciesName)) 


# Hiermee haal ik de unieke soorten eruit, die ga ik fuzzy matchen met de 
# WF dataset zodat ik nog maar een entry over hou. 
species_to_match <- last_duplicates %>% 
  distinct() %>%
  pull(AccSpeciesName)

wfo_match <- WFO.match(spec.data = species_to_match, WFO.data = classification)

matched <- wfo_match %>%
  select(spec.name, family) %>%
  rename(AccSpeciesName = spec.name, Family = family)

species_duplicates <- selected_duplicates %>%
  select(AccSpeciesName, Genus, Family) 

final_match <- inner_join(matched, species_duplicates) %>%
  distinct() %>%
  rename(family = Family, genus = Genus) 

# Deze dataset heeft enkel unieke namen. n = 67454.
final_tax_info <- bind_rows(final_match, unique_species, filter) %>%
  select(-c(AccSpeciesID, try_species)) %>%
    mutate(family = recode(family, "Agavaceae" = "Asparagaceae",
                         "Agdestidaceae" = "Phytolaccaceae",
                         "Alliaceae" = "Amaryllidaceae",
                         "Bombacaceae" = "Malvaceae",
                         "Byttneriaceae" = "Malvaceae",
                         "Compositae" = "Asteraceae",
                         "Dipsacaceae" = "Caprifoliaceae",
                         "Ehretiaceae" = "Boraginaceae",
                         "Leguminosae" = "Fabaceae",
                         "Fumariaceae" = "Papaveraceae",
                         "Hemerocallidaceae" = "Asphodelaceae",
                        "Hemidictyaceae" = "Aspleniaceae",
                        "Hyacinthaceae" = "Asparagaceae",
                         "Hydrophyllaceae" = "Boraginaceae",
                        "Laxmanniaceae" = "Asparagaceae",
                        "Malesherbiaceae" = "Passifloraceae",
                        "Myrsinaceae" = "Primulaceae",
                        "Pentapetaceae" = "Malvaceae",
                        "Petiveriaceae" = "Phytolaccaceae",
                        "Ruscaceae" = "Asparagaceae",
                        "Sambucaceae" = "Adoxaceae",
                        "Sparrmanniaceae" = "Malvaceae",
                        "Sterculiaceae" = "Malvaceae",
                        "Ternstroemiaceae" = "Pentaphylacaceae",
                        "Themidaceae" = "Asparagaceae",
                        "Thesiaceae" = "Santalaceae",
                        "Tiliaceae" = "Malvaceae",
                        "Turneraceae" = "Passifloraceae",
                        "Valerianaceae" = "Caprifoliaceae",
                        "Viburnaceae" = "Adoxaceae", 
                        "Viscaceae" = "Santalaceae",
                        "Ximeniaceae" = "Olacaceae",
                        "Cordiaceae" = "Boraginaceae")) %>%
  filter(family %notin% c("Bryaceae", "Daltoniaceae", "Leucobryaceae",
                          "Pallaviciniaceae", "Pylaisiaceae")) # Moss, liverworts families

final_species_list <- final_tax_info %>%
  pull(AccSpeciesName)
# Vraag, maakt BHPMF gebruik van ordes?

# Het doel van dit alles is om een dataset te creëren met enkel unieke soortsnamen
# en ieder met informatie over het geslacht en genus. 

hierarchy <- lookup_table(final_species_list, by_species=TRUE, 
                          family.tax = "plantlist", missing_action = "NA") %>%
 select(family, order, group)  %>%
  distinct() %>% 
  drop_na()

rownames(hierarchy) <- NULL

final_tax_info2 <- full_join(final_tax_info, hierarchy) %>%
  mutate(order =  case_when(family == "Alzateaceae" ~ "Myrtales",
                            family == "Asphodelaceae" ~ "Asparagales",
                            family == "Asteraceae" ~ "Asterales",
                            family == "Bataceae"  ~ "Brassicales",
                            family == "Crypteroniaceae" ~ "Myrtales",
                            family == "Didymochlaenaceae" ~ "Polypodiales",
                            family == "Fabaceae" ~ "Fabales",
                            family ==  "Francoaceae" ~ "Geraniales",
                            family == "Griseliniaceae" ~ "Apiales",
                            family == "Heliotropiaceae" ~ "	Boraginales",
                            family == "Isoetaceae" ~ "Isoetales",
                            family == "Kewaceae" ~ "Caryophyllales",
                            family == "Lonchitidaceae" ~ "Polypodiales",
                            family == "Macarthuriaceae" ~ "Caryophyllales",
                            family == "Mazaceae" ~ "Lamiales",
                            family == "Namaceae" ~ "Boraginales",
                            family == "Nyssaceae" ~ "Cornales",
                            family == "Ripogonaceae" ~ "Liliales",
                            family == "Thomandersiaceae" ~ "Lamiales",
                                  TRUE ~ order)) %>%
  drop_na(AccSpeciesName) %>%
  filter(order %notin% c("Bryophytes")) %>%
  select(-c(group))


try_clean_wide <- try_clean %>%
  filter(AccSpeciesName %in% c(final_tax_info2$AccSpeciesName)) %>%
  full_join(final_tax_info2) %>%
  select(-c(TraitID)) %>%
  pivot_wider(values_from = "StdValue", names_from = "TraitName") %>%
  mutate(AccSpeciesName = recode(AccSpeciesName, "Abarema langsdorffii" = "Abarema langsdorfii"))

save(try_clean_wide, 
     file = "data/processed/try_clean_wide.RData")

####### NA counting


na_test <- try_clean_wide %>%
  mutate(ObservationID = as.character(ObservationID)) %>%
  group_by(AccSpeciesName, order, family, genus) %>%
  summarise(across(where(is.numeric), ~mean(., na.rm = TRUE)))

temp <- try_clean_wide %>%
  select(-c(ObservationID))

na_sum <- colSums(is.na(temp))  %>% 
  as.data.frame() %>%
  mutate(percentage = ./nrow(temp)*100) 

na_sum2 <- colSums(is.na(na_test))  %>% 
  as.data.frame() %>%
  mutate(percentage2 = ./nrow(na_test)*100)


######################################### ROMMEL?


try_jezus <- try_filtered %>%
  select(-c(genus, sp)) %>%
  mutate(try_orig = "Yes") %>%
  full_join(try_taxonomic_2) 



length(unique(try_jezus$AccSpeciesName))

try_clean <- try_filtered %>%
  mutate(OrigValueStr =  case_when(OrigValueStr == "c3" ~ "C3",
                             OrigValueStr == "C3?" ~ "C3",
                             OrigValueStr == "c4" ~ "C4",
                             OrigValueStr == "C4?" ~ "C4",
                             TRUE ~ OrigValueStr))  %>%
  filter(DataID %notin% c(8178, 9780, 1629, 1739, 974, 2526, 2760, 2527, 2761, 
                          9906, 2762, 9578, 9577, 9576, 9575, 9574, 3710, 2958, 
                          9447, 8651, 8643, 8648, 8642, 8650, 8647, 8646, 8649, 
                          8645, 8644, 2373, 2380, 2365, 2367, 2375, 2381, 3809, 
                          3812, 3814, 3811, 3816, 225, 2082, 2142, 3990, 3991, 
                          2412, 2413, 2414, 904, 1271, 9926, 9931, 27, 9898, 
                          9899, 9900, 9901, 476, 2506, 829, 2499, 2500, 1027, 
                          1028, 2311, 4142, 3974, 3979, 3980, 3978, 3976, 3975, 
                          3977)) %>%
  filter(TraitID %in% c(3, 4, 13, 14, 15, 18, 22, 24, 26, 27, 46, 47, 48, 53, 54, 
                        59, 70, 71, 83, 89, 145, 146, 164, 246, 334, 785, 819, 839, 
                        1047, 1080, 1181, 1256, 3106, 3113, 3117, 3588, 3801)) %>%
  drop_na(StdValue) %>%
  select(AccSpeciesName, TraitID, ObservationID, TraitName, StdValue) %>%
  group_by(AccSpeciesName, ObservationID, TraitID, TraitName) %>%
  summarise(across(where(is.numeric), ~mean(., na.rm = TRUE))) %>%
  ungroup() %>%
  
# This extracts all the species names from the try database
try_species <- try_clean %>%
  select(AccSpeciesName) %>%
  distinct() %>%
  mutate(try_species = "Yes") # 68812 species in total

# The current, Accepted Species Name list is hold againts the classification file from
# the World Flora database to check if there is a perfect match. 
try_tax_wf_first <- classification %>%
  filter(taxonomicStatus != "Synonym") %>%
  filter(taxonID %notin% c("wfo-1200010625", "wfo-1200010660", "wfo-0001233171")) %>% 
  # Remove entries of three species with the wrong family name 
  select(scientificName, family, genus) %>%
  rename(AccSpeciesName = scientificName) %>%
  left_join(try_species) %>%
  distinct() %>%
  filter(try_species == "Yes") #57379 species are present in the World Flora database

# Not all species have a perfect match, this retrieves the species names that are not in the WorldFlora database
not_included <- try_species %>%
  filter(AccSpeciesName %notin% c(try_tax_wf_first$AccSpeciesName)) %>%
  pull(AccSpeciesName)

# They are fuzzy matched against the WFO database. 
wfo_match <- WFO.match(spec.data = not_included, WFO.data = classification)

# The species that have a match now
included_now <- wfo_match %>%
  filter(Matched == TRUE) %>%
  filter(taxonRank %notin% c("genus", "family")) %>%
  separate(spec.name, into = c("genus", "sp"), extra = "merge", remove = FALSE) %>%
  filter(sp != "sp") %>%
  filter(Subseq == 1)


# Species that don't have a match
still_not_included <- tax_test %>%
  filter(Matched == FALSE) %>%
  separate(spec.name, into = c("genus", "sp"), extra = "merge", remove = FALSE) %>%
  filter(sp != "sp")  %>%
  mutate(spec.name_new = tolower(spec.name)) %>%
  mutate(spec.name_new = paste0(tools::toTitleCase(word(spec.name_new, 1)), " ", word(spec.name_new, 2, -1))) %>%
  select(spec.name, spec.name_new) %>%
  rename(spec.name_old = spec.name, spec.name = spec.name_new)


still_not_included_list <- still_not_included %>%
  pull(spec.name)

tax_test2 <- WFO.match(spec.data = still_not_included_list, WFO.data = classification) 

included_now_2 <- tax_test2 %>%
  filter(taxonRank %notin% c("genus", "family")) %>%
  filter(Matched == TRUE) %>%
  filter(Subseq == 1) 

included_now_both <- bind_rows(included_now, included_now_2) %>%
  select(spec.name, scientificName) %>%
  rename(AccSpeciesName = spec.name, new_name = scientificName) %>%
  filter(AccSpeciesName %notin% c( "Dry alluvial", "Serpentine grass", 
  "Perennial grasses", "deciduous chaparral shrubs", "Mesophytic meadow",
  "tropical evergreen trees", "Graminoid abandoned area", "Harvest Residues",
  "Pubescent rosette", "Senna form taxon 'filifolia'", "Stoloniferous grass",
  "Graminoid pasture", "Meadow steppe", "Tropical Cultivar", "Dwarf shrub", "Graminoid meadow", 
  "Tropical manaus",  "Wetland grass", "Annual grass",  "Fava de", "Para para", "Ita alba preta",
  "Eri sch", "Desmopsis laeve", "Sloanea oppd", "Carex dark green", "Envira preta", 
  "Randia gorky", "Bunchosia myrt", "Castanha de", "Cupania verde", "Zombia anomala",
  "Sloanea oak", "Abacaba palm", "Araca da mata", "Cacao da mata", "Tall catinga species",
  "Ucuuba da terra firme", "Ucuuba folha peluda", "Ucuuba t folha", "Ucuuba vermelha")) %>%
  mutate(try_species = "Yes") 
# Some found matches are not plausible. And some should not have been matched at all.
# These are filtered out. 


try_clean_wide <- try_clean %>%
  select(-c(TraitID)) %>%
  pivot_wider(values_from = StdValue, names_from = TraitName) %>%
  mutate(AccSpeciesName = tolower(AccSpeciesName)) %>%
  mutate(AccSpeciesName = paste0(tools::toTitleCase(word(AccSpeciesName, 1)), " ", word(AccSpeciesName, 2, -1))) 

try_clean_wide_1 <- try_clean_wide %>%
  filter(AccSpeciesName %in% c(try_tax_wf$AccSpeciesName)) 

try_clean_wide_2 <- included_now_both %>%
  left_join(try_clean_wide) %>%
  select(-c(AccSpeciesName)) %>%
  rename(AccSpeciesName = new_name) 

try_clean_wide_both <- bind_rows(try_clean_wide_1, try_clean_wide_2) %>%
  select(-c(new_name, try_species)) %>%
  mutate(try_species = "Yes")


try_tax_wf_final <- classification %>%
  filter(taxonomicStatus != "Synonym") %>%
  filter(taxonID %notin% c("wfo-1200010625", "wfo-1200010660", "wfo-0001233171")) %>% 
  # Remove entries of three species with the wrong family name 
  select(scientificName, family, genus) %>%
  rename(AccSpeciesName = scientificName) %>%
  distinct() %>%
  left_join(try_clean_wide_both) %>%
  filter(try_species == "Yes") %>%
  select(-c(try_species))

na_sum <- colSums(is.na(try_tax_wf_final))  %>% 
  as.data.frame() %>%
  mutate(percentage = ./66782*100)



save(try_filtered,
  file = "data/processed/try_filtered2.RData")

save(try_tax_wf_final, 
     file = "data/processed/try_tax.RData")


temp <- try_filtered %>%
  filter(ObservationID == 3061340)


```

```{r BHPMF}

trait <- LEDA_all %>%
  select(-c(genus, species, TraitID, o)) %>%
  column_to_rownames(var = "Name") %>%
  as.matrix()



#The data must be normalized and linearized. A log transformation is used. But it could be that that is not the appropriate transformation. 
back_trans_pars <- list()
rm_col <- c()

for(i in 1:ncol(trait)){
x <- trait[,i] # goes through the columns
min_x <- min(x,na.rm = T) # takes the min of each column
if(min_x < 0.00000000001){
x <- x - min_x + 1 # make this optional if min x is neg
}
logx <- log10(x)
mlogx <- mean(logx, na.rm = T)
slogx <- sd(logx, na.rm = T)
x <- (logx - mlogx)/slogx # Z transformation
back_trans_pars[[i]] <- list(min_x = min_x,
mlogx = mlogx,
slogx = slogx)
trait[,i] <- x
}





```

```{r rommel?}

try_filtered <- try_filtered %>%
  filter(AccSpeciesName %notin% c("unknown", "Herbaceous", "Graminoid", "grass", "Buxuxu", "Palm", "Tree", 
  "Bamboo", "Shrub", "Vine", "forb", "Fern", "Pallida", "Birch", "Spruce", "Triticosecale", 
  "Tachi", "Taxi", "Mid", "Geophyte", "Grasses", "Conifers", "Forest", "Muiraba", "Ucuubarana", 
  "Jutai-mirim", "Amapai", "Amarelao", "Muiracatiara", "Janita", "Jutai-acu", "Muirapinima", 
  "Jarana", "Louro", "Morototo", "Muirauba", "GUARUBARANA", "AROEIRA", "GUARIUBA", "MAMORANA", 
  "URUCURANA", "QUINARANA", "CAFERANA", "ANARAI", "PITOMBA", "CHICUA", "POIZOROQUEIRA", 
  "Quarubarana", "ANANI", "GOIABINHA", "ABIURANA", "PIQUIA", "ACHICHA", "SABOEIRO", "COCAO", 
  "PARAPARA", "SUCUUBA", "URUAZEIRO", "MUIRAPICHONA", "AMADAI", "EMBAUBARANA", "INAJARANA", 
  "MUIRATINGA", "LIMORANA", "ANDIROBARANA", "BABACU", "MUIRARANA", "CUNARIO", "GINJA", 
  "ACARIQUARANA", "ARATAJILRANA", "TATAPIRIRICA", "QUARIQUARANA", "ACARIQUARA", "PUROI", 
  "PARA", "ANDIROBA", "CARALLE", "ARATASIURANA", "PAMA", "CARAPAMAUBA", "PAPAROLA", "CARADANAUBA", 
  "ROSADINHO", "JATAUBA", "BARBATINAO", "CUPUI", "CIIJIA", "SARDINHEIRA", "PASACINHEIRA", "AQUIQUI", 
  "AMARELINHO", "CUNAREO", "CUNARE", "JINGA", "MACUCU", "EMBAUBA", "MURUPITA", "PITAICA", "PADAROLA", 
  "CAJUI", "QUARIQUARA", "MUTUTI", "JENIPARANA", "INAJA", "AMAVAI", "GOIABARNA", "ABDILOBARANA", 
  "MUIRARENGA", "PASARINHARA", "PATIA", "MUIRAPIRANGA", "ACHUA", "Pajurazinho", "QUATAQUACAUA", 
  "BACABA", "MUUBA", "CUMAT", "BABTIMAO", "MAPARAJUBA", "UCHIRANA", "Bindweed", "Triticale", 
  "Brassolaeliocattleya_x", "Moss")) 




```
