==> src/old_scripts/3. Complete_castro.Rmd <==
---
title: "0.Preprocessing"
author: "Jelyn Gerkema"
date: "2023-06-20"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 Castro et al. 2010
```{r clean data}

castro_species_list <- read_excel("data/raw/CESTES/Plant_data/Castro2010_AJ.xlsx", 
    sheet = "splist") %>%
  mutate(Taxon = recode(Taxon, "Hypochoeris glabra" = "Hypochaeris glabra"))


plot_list <- read_excel("data/raw/CESTES/Plant_data/Castro2010_AJ.xlsx", 
    sheet = "sitelist") %>%
    rename(Plot_code = Sites, 
           Sites = SiteID) %>%
    mutate(across(.cols = Sites, as.character))

castro_veg <- read_excel("data/raw/CESTES/Plant_data/Castro2010_AJ.xlsx") %>%
  drop_na() %>%
  column_to_rownames(var = "Sites")

colnames(castro_veg) <- castro_species_list$Taxon

castro_veg <- castro_veg %>%
  rownames_to_column(var = "Sites") %>%
  full_join(plot_list) %>%
  relocate(Plot_code) %>%
  separate(Plot_code, into = c("Treatment", "Replicate"), remove = F, sep = "(?<=\\D)(?=\\d)|(?<=\\d)(?=\\D)") 

castro_vegtemp <- castro_veg %>%
  select(-c(Sites, Replicate)) %>%
  pivot_longer(cols = 3:ncol(.), names_to = "Name", values_to = "cover")

castro_veg <- castro_veg %>%
  select(-c(Sites, Replicate, Treatment)) %>%
  column_to_rownames(var = "Plot_code") %>% 
  as.matrix()


# They did not measure traits for each species in each species. Instead, they measured the traits only for the 80% most abundant ones. 

castro_trait <- read_excel("data/raw/CESTES/Plant_data/Castro2010_AJ.xlsx", 
    sheet = "traitsfull") %>%
  separate(Spraw, into = c("Name", "Treatment"), sep = "_") %>%
  rename(Height = VH) %>%
  select(-c(LF, M, Pollination, Dispersal, LPC, Spnew, Sp)) %>% # LPC removed, was not used in the original publication'
  mutate(Name = recode(Name, "Hypochoeris glabra" = "Hypochaeris glabra")) %>%
  full_join(castro_vegtemp) %>%
  group_by(Name) %>%
  mutate(across(where(is.numeric), ~replace_na(., 
                        mean(., na.rm = TRUE)))) %>% # The NA's, caused by the absence of species in certain plots, are filled up by the species mean
  ungroup() %>%
  select(-c(cover, Treatment)) %>%
  select(order(colnames(.))) %>%
  relocate(Plot_code, Name) %>%
  arrange(Name) 

castro_trait_fd <- castro_trait 

# Standardizing is done after the fd dataframe is made, otherwise the values are standardized twice
castro_trait <- castro_trait %>%
  mutate(across(where(is.numeric), ~ c(scale(., center = F)))) 


traitlist <- sort(c("Height", "LCC", "LDMC", "LNC", "SLA", "SM"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)




``` 


## 5. Perform calculations
```{r calculations }

castro_processed_specieslvl <- cats_preparation(input_veg = castro_veg, input_trait = castro_trait,
                                      cutoff_value =  0.8, trait_information = trait_information,
                                      trait_detail = "Treatment")

mean_dist <- calculate_mean_dist(castro_veg, castro_processed_specieslvl)

FD_df <- calculate_FD_plotlvl(castro_processed_specieslvl, castro_veg, castro_trait_fd)

diversity_info <- castro_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]] %>% 
  as.data.frame() %>% 
  mutate(shannon = diversity(.[,1:ncol(.)], index = "shannon", MARGIN = 1, base = exp(1)),
         simpson = diversity(.[,1:ncol(.)], index = "simpson", MARGIN = 1, base =exp(1)),
         species_richness = specnumber(.[,1:ncol(.)], MARGIN = 1),
         empty_states = ncol(castro_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]]) - species_richness,
         meta_size = ncol(castro_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]])) %>%
  select(shannon, simpson, species_richness, empty_states, meta_size) %>% # base = exp(1) gives the shannon index a log e base
  rownames_to_column(var = "Plot_code") %>%
  full_join(mean_dist) %>%
  full_join(FD_df)

```



## 6. Cats model
```{r cats calculations}
#Determines amount of cores, substract four to keep to pc free for other processes.
n_cores <- parallel::detectCores() - 4 

my_cluster <- makeSOCKcluster(n_cores)

registerDoSNOW(cl = my_cluster)

foreach::getDoParRegistered() # check if cluster is set up

#########################

castro_cats <- cats_calculations(castro_processed_specieslvl, trait_information,
                                 trait_detail = "Treatment") %>%
  mutate(study = "Castro")

castro_figure <- cats_visualisation(castro_cats) %>%
  mutate(study = "Castro")

save(castro_processed_specieslvl, castro_figure, castro_species_list, castro_cats,
  file = "data/processed/castro_calculations.RData")

castro_figure <- cats_visualisation(castro_cats) %>%
  separate(Plot_code, into = c("Treatment", "Replicate"), remove = F, sep = "(?<=\\D)(?=\\d)|(?<=\\d)(?=\\D)") 



ggplot(castro_figure, aes(x = species_richness, y = KLR2)) +
  geom_point(aes(colour = Treatment)) +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm") + 
  labs(title = "The relationship between species richness and the CATS decomposition - Castro et al. 2010",
       x = "Species richness") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.6, 1)) +
  scale_y_continuous(breaks = c(-0.6, -0.4, -0.2, 0, 0.2, 0.4, 
    0.6,  0.8,  1
  ))

ggplot(castro_figure, aes(x = mean_dist, y = KLR2)) +
  geom_point(aes(colour = Treatment)) +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm") + 
  labs(title = "The relationship between plot dissimilarity and the CATS decomposition - Castro et al. 2010",  x = "Average Bray-Curtis index value") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.6, 1)) +
  scale_y_continuous(breaks = c(-0.6, -0.4, -0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  )) 



```
