==> src/old_scripts/3. Complete_chalmandrier.rmd <==
---
title: "0.Preprocessing"
author: "Jelyn Gerkema"
date: "2023-06-20"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")
```

## 1.Clean data

### 1.1 Chalmandrier et al. 2022
```{r clean data}

# The data of Chalmandrier has already been filtered in their own analysis. 
# I use their raw cover and trait data but I extract their selected plot_codes so I can filter the raw data

load("data/raw/Chalmandrier et al. 2022/data/data_ready.Rdata")

plot_code_vector <- env_p %>%
  pull(QuadID)

# Remove unnecessary objects from the global environment
rm(data)
rm(obs.comm)
rm(pdf_species)
rm(t.avg)
rm(env_p)

# Cleaning the vegetation data so it can be used to filter the trait data. 
# Some plots have traits but no cover data. The traits are used to impute missing values
# but the plots are afterwards removed. Some plots have cover but no traits, these traits
# are imputed by taking the species average across all plots. Some plots have cover and multiple
# trait values for some species. For those plots, the average species value is taken. In addition,
# trait data is created for the empty states,  ie. absent species in a certain plot. 

chalmandrier_species_list <- read_excel("data/raw/Chalmandrier et al. 2022/data/complete_species_names.xlsx")

chalmandrier_veg_cleantemp <- read_csv("data/raw/Chalmandrier et al. 2022/data/community.cover.23sp.csv") %>%
  filter(QuadID %in% plot_code_vector) %>%
  rename(Plot_code = QuadID) %>%
  pivot_longer(cols = 2:ncol(.), values_to = "cover", names_to = "Six-letter code") %>%
  left_join(chalmandrier_species_list) %>%
  select(-c(`Six-letter code`)) %>%
  rename(Name = Species)

chalmandrier_trait <- read_csv("data/raw/Chalmandrier et al. 2022/data/kettleholedata.23sp.csv") %>%
  filter(QuadID %in% plot_code_vector) %>%
  select(QuadID, poros, height, sla, species) %>%
  rename(Plot_code = QuadID, `Six-letter code` = species) %>%
  left_join(chalmandrier_species_list) %>%
  rename(Name = Species) %>%
  select(-c(`Six-letter code`)) %>%
  group_by(Plot_code, Name) %>%
  summarise(across(where(is.numeric), ~mean(., na.rm = TRUE))) %>% #In some quadrants, some species have multiple recordings, the average is taken here. 
  ungroup() %>%
  full_join(chalmandrier_veg_cleantemp) %>%
  group_by(Name) %>%
  mutate(across(.cols = c(poros, height, sla),  
            ~replace_na(., 
                        mean(., na.rm = TRUE)))) %>% # The NA's are filled up by the species mean
  ungroup() %>%
  arrange(Name) %>%
  drop_na(cover) %>%
  drop_na(poros) %>% # Some cases have no abundance data and no trait data, by dropping NA's 
  # of one of the trait values, these cases (n = 134) are removed as well. 
  select(-c(cover)) %>%
  select(order(colnames(.))) %>%
  relocate(Plot_code, Name)

chalmandrier_trait_fd <- chalmandrier_trait 

# Standardizing is done after the fd dataframe is made, otherwise the values are standardized twice
chalmandrier_trait <- chalmandrier_trait %>%
  mutate(across(where(is.numeric), ~ c(scale(., center = F)))) 


traitlist <- sort(c("height", "poros", "sla"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)

# Cleaning the vegetation data so it can be used in the calculate_abundance function
chalmandrier_veg <- chalmandrier_veg_cleantemp %>%
  pivot_wider(names_from = "Name", values_from = "cover") %>%
  column_to_rownames(var = "Plot_code") %>%
  select(order(colnames(.)))


``` 

# Calculations
```{r chalmandrier calculations}

chalmandrier_processed_specieslvl <- cats_preparation(input_veg = chalmandrier_veg, input_trait = chalmandrier_trait,  
                                                  trait_information, 0.8, trait_detail = "Plot") 

mean_dist <- calculate_mean_dist(chalmandrier_veg, chalmandrier_processed_specieslvl)

FD_df <- calculate_FD_plotlvl(chalmandrier_processed_specieslvl, chalmandrier_veg, chalmandrier_trait_fd)


diversity_info <- chalmandrier_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]] %>% 
  as.data.frame() %>% 
  mutate(shannon = diversity(.[,1:ncol(.)], index = "shannon", MARGIN = 1, base = exp(1)),
         simpson = diversity(.[,1:ncol(.)], index = "simpson", MARGIN = 1, base =exp(1)),
         species_richness = specnumber(.[,1:ncol(.)], MARGIN = 1),
         empty_states = ncol(chalmandrier_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]]) - species_richness,
         meta_size = ncol(chalmandrier_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]])) %>%
  select(shannon, simpson, species_richness, empty_states, meta_size) %>% # base = exp(1) gives the shannon index a log e base
  rownames_to_column(var = "Plot_code") %>%
  full_join(mean_dist) %>%
  full_join(FD_df)



```

## 6. CATS calculations
```{r Cats calculations}
#Determines amount of cores, substract four to keep to pc free for other processes.
n_cores <- parallel::detectCores() - 4 

my_cluster <- makeSOCKcluster(n_cores)

registerDoSNOW(cl = my_cluster)

foreach::getDoParRegistered() # check if cluster is set up

#####

chalmandrier_cats <- cats_calculations(chalmandrier_processed_specieslvl, trait_information, trait_detail = "Plot") %>%
  mutate(study = "Chalmandrier")

chalmandrier_figure <- cats_visualisation(chalmandrier_cats) %>%
  mutate(study = "Chalmandrier")

save(chalmandrier_processed_specieslvl, chalmandrier_figure, chalmandrier_species_list, chalmandrier_cats,
  file = "data/processed/chalmandrier_calculations.RData")

ggplot(chalmandrier_figure, aes(x = species_richness, y = KLR2)) +
  geom_point() +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm") + 
  labs(title = "The relationship between species richness and the CATS decomposition - Chalmandrier et al. 2022",
       x = "Species richness") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.4, 1)) +
  scale_y_continuous(breaks = c(-0.4, -0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  ))

ggplot(chalmandrier_figure, aes(x = mean_dist, y = KLR2)) +
  geom_point() +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm") + 
  labs(title = "The relationship between plot dissimilarity and the CATS decomposition - Chalmandrier et al. 2022",  x = "Average Bray-Curtis index value") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.4, 1)) +
  scale_y_continuous(breaks = c(-0.4, -0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  )) 


```
