==> src/old_scripts/3. Complete_delariva.rmd <==
---
title: "0.Preprocessing"
author: "Jelyn Gerkema"
date: "2023-06-20"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 De la Riva et al. 2015
```{r clean data}

riva_veg_cleantemp <- read_excel("data/raw/de la Riva et al. 2015/veg_data.xlsx") %>%
  mutate_all(na_if, "-") %>%
  mutate(across(Virgen_Upper:Villaviciosa_Lower, as.numeric)) %>%
  replace(is.na(.), 0) %>%
  select(-c(Family, `Life habit`, `Leaf habit`)) %>%
  arrange(Species) %>%
  pivot_longer(cols = 2:ncol(.), values_to = "cover", names_to = "Plot_code") %>%
  rename(Name = Species) %>%
  mutate(Name = recode(Name, "Cistus salvifolius" = "Cistus salviifolius" ))


riva_veg <- read_excel("data/raw/de la Riva et al. 2015/veg_data.xlsx") %>%
  mutate_all(na_if, "-") %>%
  mutate(across(Virgen_Upper:Villaviciosa_Lower, as.numeric)) %>%
  replace(is.na(.), 0) %>%
  select(-c(Family, `Life habit`, `Leaf habit`)) %>%
  arrange(Species) %>%
  mutate(Species = recode(Species, "Cistus salvifolius" = "Cistus salviifolius" )) %>%
  column_to_rownames(var = "Species") %>%
  t()


riva_trait <- read_excel("data/raw/de la Riva et al. 2015/Dryad_database.xls") %>%
  rename(Height = `Plant height (m)`,
         Leaf_size = `Leaf size  (cm²)`,
         LDMC = `LDMC        (g g¯¹)`,
         SLA = `SLA              (m² Kg¯¹)`,
         LChl = `LChl (µg g¯¹)`,
         LNC = `LNC (%)`,
         d13C = `δ 13C (‰)`,
         SDMC = `SDMC       (g g¯¹)`,
         WD = `WD                  (g cm¯³)`,
         RDMC = `RDMC         (g g¯¹)`,
         SRL = `SRL (m g¯¹)`,
         Name = Species) %>%
  mutate(Zone = recode(Zone, `Virgen Linares` = "Virgen")) %>%
  mutate(Plot_code = paste0(Zone, "_", Plot)) %>%
  select(-c(Zone, Plot, Country, Province, `Plant cover (m²)`, 
            Latitude, Longitude, d13C)) %>%
  select(order(colnames(.))) %>%
  relocate(Plot_code, Name) %>%
  mutate(Name = recode(Name, `Arbutus unedu` = "Arbutus unedo",
                       `Myrtus Communis` = "Myrtus communis",
                       `Pistacia letiscus`  = "Pistacia lentiscus",
                       `Quecus cocifera`  = "Quercus coccifera",
                       `Rubus ulmilifolius` = "Rubus ulmifolius",
                       `Rubus umilifolius` = "Rubus ulmifolius",
                       `Crataegus monogina`  = "Crataegus monogyna",
                       `Phyllirea angustifolia` = "Phillyrea angustifolia",
                       `Phyllirea latifolia` = "Phillyrea latifolia",
                       `Pyrus bourgeana` = "Pyrus bourgaeana",
                       `Cistus salvifolius` = "Cistus salviifolius")) %>%
  full_join(riva_veg_cleantemp) %>%
  group_by(Name) %>%
  mutate(across(where(is.numeric), ~replace_na(., mean(., na.rm = TRUE)))) %>%
  arrange(Name) %>%
  select(-c(cover)) %>%
  ungroup()


riva_trait_fd <- riva_trait 

# Standardizing is done after the fd dataframe is made, otherwise the values are standardized twice
riva_trait <- riva_trait %>%
  mutate(across(where(is.numeric), ~ c(scale(., center = F)))) 


riva_species_list <- colnames(riva_veg)

traitlist <- sort(c("Height", "Leaf_size",  "LDMC", "SLA", "LChl", "LNC", 
                "SDMC", "WD", "RDMC", "SRL"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)

``` 


## 5. Perform calculations
```{r calculations }

riva_processed_specieslvl <-  cats_preparation(riva_veg, riva_trait, trait_information, 0.8,
                                               trait_detail = "Plot")

mean_dist <- calculate_mean_dist(riva_veg, riva_processed_specieslvl)

FD_df <- calculate_FD_plotlvl(riva_processed_specieslvl, riva_veg, riva_trait_fd)


diversity_info <- riva_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]] %>% 
  as.data.frame() %>% 
  mutate(shannon = diversity(.[,1:ncol(.)], index = "shannon", MARGIN = 1, base = exp(1)),
         simpson = diversity(.[,1:ncol(.)], index = "simpson", MARGIN = 1, base =exp(1)),
         species_richness = specnumber(.[,1:ncol(.)], MARGIN = 1),
         empty_states = ncol(riva_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]]) - species_richness,
         meta_size = ncol(riva_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]])) %>%
  select(shannon, simpson, species_richness, empty_states, meta_size) %>% # base = exp(1) gives the shannon index a log e base
  rownames_to_column(var = "Plot_code") %>%
  full_join(mean_dist) %>%
  full_join(FD_df)


```

## 6. Run the CATS model and visualize results
```{r CATS model}
#Determines amount of cores, substract four to keep to pc free for other processes.
n_cores <- parallel::detectCores() - 4 

my_cluster <- makeSOCKcluster(n_cores)

registerDoSNOW(cl = my_cluster)

foreach::getDoParRegistered() # check if cluster is set up


riva_cats <- cats_calculations(riva_processed_specieslvl, trait_information,
                          trait_detail = "Plot") %>%
  mutate(study = "Riva")

riva_figure <- cats_visualisation(riva_cats) %>%
  mutate(study = "Riva")

save(riva_processed_specieslvl, riva_figure, riva_species_list, riva_cats,
  file = "data/processed/riva_calculations.RData")


ggplot(riva_figure, aes(x = species_richness, y = KLR2)) +
  geom_point() +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm") + 
  labs(title = "The relationship between species richness and the CATS decomposition - de la Riva et al. 2015",
       x = "Species richness") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.2, 1)) +
  scale_y_continuous(breaks = c(-0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  ))


ggplot(riva_figure, aes(x = mean_dist, y = KLR2)) +
  geom_point() +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm") + 
  labs(title = "The relationship between plot dissimilarity and the CATS decomposition - de la Riva et al. 2015",  x = "Average Bray-Curtis index value") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.2, 1)) +
  scale_y_continuous(breaks = c(-0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  )) 


```
