==> src/old_scripts/3. Complete_frenette_2010.Rmd <==
---
title: "3. Complete analysis Eallonardo et al."
author: "Jelyn Gerkema"
date: "2023-06-20"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 Frenette-Dussault et al. 2012
```{r clean data}

frenette2010_veg_info <- read_excel("data/raw/CESTES/Plant_data/Frenette2012b_AJ.xlsx", 
    sheet = "envirfull") %>%
    rename(Plot_code = Sites) %>%
  mutate(across(Plot_code, as.character))

frenette_veg <- read_excel("data/raw/CESTES/Plant_data/Frenette2012b_AJ.xlsx", 
    sheet = "commfull") %>%
    column_to_rownames(var = "Sites" ) 

frenette2010_species_list <- read_excel("data/raw/CESTES/Plant_data/Frenette2012b_AJ.xlsx", 
    sheet = "splist") %>%
  rename(Sp = TaxCode, Name = Taxon) %>%
  mutate(Name = recode(Name, "Helianthemum lipii" = "Helianthemum lippii",
                       "Salsola verticilata" = "Salsola verticillata"))


colnames(frenette_veg) <- frenette2010_species_list$Name

frenette_veg <- frenette_veg %>%
    select(where(~ any(. != 0)))  # One species is not present anywhere and is filtered out through this
    
species_list_temp <- frenette2010_species_list %>% pull(Name) 

hierarchy <- lookup_table(species_list_temp, by_species=TRUE, 
                          family.tax = "plantlist", missing_action = "drop") %>%
  rownames_to_column(var = "Name")

family_mean <- read_excel("data/raw/CESTES/Plant_data/Frenette2012b_AJ.xlsx", 
  sheet = "traitsfull")  %>%
  full_join(frenette2010_species_list) %>%
  full_join(hierarchy) %>%
  group_by(family) %>%
  mutate(across(MH, as.numeric)) %>%
  summarise(across(where(is.numeric), ~mean(., na.rm = T))) %>%
  filter(family == "Leguminosae") %>%
  pull(MH) 

frenette_trait <- read_excel("data/raw/CESTES/Plant_data/Frenette2012b_AJ.xlsx", 
  sheet = "traitsfull") %>%
  full_join(frenette2010_species_list) %>% 
  select(-c(LC13, LN15, SM, GrowthForm, FlowerOnset, Clon, VH, PosInflor, Sp,
            Succ, PastoralVal, LCNR)) %>%
  rename(C4 = PhotoPath,
         Spinescence = Spin, Height = MH) %>%
  mutate(C4 = recode(C4, `C4` = 2,
                            `C3`  =  1),
         Spinescence = recode(Spinescence, `Presence` = 2,
                             `Absence`  =  1)) %>%
  mutate(across(Height, as.numeric)) %>%
  replace_na(list(Height = family_mean)) %>%
  select(order(colnames(.))) %>%
  relocate(Name) %>%
  arrange(Name) %>%
  filter(Name %in% colnames(frenette_veg)) %>%
  mutate(across(where(is.numeric), ~ c(scale(., center = FALSE)))) 

frenette_trait_fd <- read_excel("data/raw/CESTES/Plant_data/Frenette2012b_AJ.xlsx", 
  sheet = "traitsfull") %>%
  full_join(frenette2010_species_list) %>% 
  select(-c(LC13, LN15, SM, GrowthForm, FlowerOnset, Clon, VH, PosInflor, Sp,
            Succ, PastoralVal, LCNR)) %>%
  rename(C4 = PhotoPath, 
         Spinescence = Spin, Height = MH) %>%
  mutate(C4 = recode(C4, `C4` = 1,
                            `C3`  =  0),
         Spinescence = recode(Spinescence, `Presence` = 1,
                             `Absence`  =  0)) %>%
  mutate(across(Height, as.numeric)) %>%
  replace_na(list(Height = family_mean)) %>%
  select(order(colnames(.))) %>%
  relocate(Name) %>%
  arrange(Name) %>%
  filter(Name %in% colnames(frenette_veg)) %>%
  column_to_rownames(var = "Name")


traitlist <- sort(c("C4", "Height", "LA", "LCC", "LDMC",
                    "LNC", "SLA", "Spinescence"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)


``` 


## 5. Perform calculations
```{r calculations }

frenette2010_processed_specieslvl <- cats_preparation(input_veg = frenette_veg, input_trait = frenette_trait, cutoff_value =  0.8, trait_information = trait_information,
                                                   trait_detail = "Study")

mean_dist <- calculate_mean_dist(frenette_veg, frenette2010_processed_specieslvl)

FD <- dbFD(frenette_trait_fd, frenette_veg, stand.x = T)

FD_df <- data.frame(FEve = FD$FEve,
        FDiv  = FD$FDiv,
        FRic = FD$FRic,
        RaoQ = FD$RaoQ,
        qual.FRic = FD$qual.FRic) %>%
  rownames_to_column(var = "Plot_code")

diversity_info <- frenette2010_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]] %>% 
  as.data.frame() %>% 
  mutate(shannon = diversity(.[,1:ncol(.)], index = "shannon", MARGIN = 1, base = exp(1)),
         simpson = diversity(.[,1:ncol(.)], index = "simpson", MARGIN = 1, base =exp(1)),
         species_richness = specnumber(.[,1:ncol(.)], MARGIN = 1),
         empty_states = ncol(frenette2010_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]]) - species_richness,
         meta_size = ncol(frenette2010_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]])) %>%
  select(shannon, simpson, species_richness, empty_states, meta_size) %>% # base = exp(1) gives the shannon index a log e base
  rownames_to_column(var = "Plot_code") %>%
  full_join(mean_dist) %>%
  full_join(FD_df)

```



## 6. Cats model
```{r cats calculations}

frenette2010_cats <- cats_calculations(frenette2010_processed_specieslvl, trait_information,
                                       trait_detail = "Study") %>%
  mutate(study = "Frenette2010")


frenette2010_figure <- cats_visualisation(frenette2010_cats) %>%
  mutate(study = "Frenette2010")

save(frenette2010_processed_specieslvl, frenette2010_figure, frenette2010_species_list,
frenette2010_veg_info, frenette2010_cats,
  file = "data/processed/frenette2010_calculations.RData")

frenette2010_figure <- frenette2010_figure %>%
  full_join(frenette2010_veg_info) 

ggplot(frenette2010_figure, aes(x = species_richness, y = KLR2)) +
  geom_point(aes(colour = Grazing)) +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm") + 
  labs(title = "The relationship between species richness and the CATS decomposition - Frenette et al., 2012 (2010)",
       x = "Species richness") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.2, 1)) +
  scale_y_continuous(breaks = c(-0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  )) +
  scale_x_continuous(breaks = c(0, 2, 4, 6, 8, 10, 12, 14))

ggplot(frenette2010_figure, aes(x = mean_dist, y = KLR2)) +
  geom_point(aes(colour = Grazing)) +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm") + 
  labs(title = "The relationship between plot dissimilarity and the CATS decomposition - Frenette et al., 2012 (2010)",
       x = "Average Bray-Curtis index value") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.2, 1)) +
  scale_y_continuous(breaks = c(-0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  )) 

```
