==> src/old_scripts/3. Complete_gonzalez.Rmd <==
---
title: "3. Complete analysis Kassaw et al."
author: "Jelyn Gerkema"
date: "2023-10-16"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 
```{r clean data}

gonzalez_species_list <- read_csv("data/raw/González et al. 2021/BIOVERA-tree_scientific_name.csv")

removed_author <- gonzalez_species_list %>%
  separate(treeIdentification, into = c("genus", "sp"), sep = " ",
           extra = "merge", remove = F ) %>%
  filter(genus %notin% c("Unidentified")) %>%
  filter(sp %notin% c("sp. 1", "sp.", "sp. 2", "sp. 3",
                      "sp. 4", "sp. 6", "sp.2")) %>%
  separate(sp, into = c("sp2", "author"), extra = "merge", sep = " ",
           remove = F) %>%
  mutate(treeIdentification = paste0(genus, " ", sp2)) %>%
  select(treeIdentification, treeIdentificationID, family) %>%
  mutate(treeIdentification = recode(treeIdentification, "Diospyros digyna" = "Diospyros nigra",
                                     "Sapium nitidum" = "Sapium lateriflorum"))
  # The synonym of Diospyros digyna Jacq. is Diospyros nigra 
 # (according to WFO: https://www.worldfloraonline.org/taxon/wfo-0000648775)

gonzalez_species_list <- gonzalez_species_list %>%
  filter(treeIdentificationID %notin% c(removed_author$treeIdentificationID)) %>%
  bind_rows(removed_author) %>%
  mutate(treeIdentificationID = recode(treeIdentificationID, "Drymis sp. 1" = "Drimys sp. 1",
                                       "Buddleia sp. 1" = "Buddleja sp. 1"))


gonzalez_veg <- read_csv("data/raw/González et al. 2021/BIOVERA-tree_community_matrix.csv") %>%
  rename(Plot_code = locationID) %>%
  column_to_rownames(var = "Plot_code") %>%
  t() %>%
  as.data.frame() %>%
  rownames_to_column(var = "treeIdentificationID") %>%
  full_join(gonzalez_species_list) %>%
  column_to_rownames(var = "treeIdentification") %>%
  select(-c(treeIdentificationID, family)) %>%
  t() %>%
  as.data.frame() %>%
  select(order(colnames(.)))  

gonzalez_trait <- read_csv("data/raw/González et al. 2021/BIOVERA-tree_functional_traits.csv") %>%
  full_join(gonzalez_species_list) %>%
  select(c(treeIdentification, traitName, traitValue)) %>%
  rename(Name = treeIdentification) %>%
  group_by(Name, traitName) %>%
  summarise(across(where(is.numeric), ~mean(., na.rm = TRUE))) %>%
  ungroup() %>%
  pivot_wider(values_from = traitValue, names_from = traitName) %>%
  drop_na() %>%
  rename(LA = leafArea, LDMC = leafDryMatterContent, Height = maximumHeight,
         SLA = specificLeafArea, WD = woodDensity) %>%
  select(order(colnames(.))) %>%
  relocate(Name) %>%
  arrange(Name)

gonzalez_trait_fd <- gonzalez_trait %>%
  column_to_rownames(var = "Name") 

species_list_fd <- gonzalez_trait %>% pull(Name)

gonzalez_veg_fd <- gonzalez_veg %>%
  select(all_of(species_list_fd))

gonzalez_trait <- gonzalez_trait %>%
  mutate(across(where(is.numeric), ~ c(scale(., center = F)))) 

traitlist <- sort(c( "chlorophyllContent", "Height", "LA", "laminaDensity", "LDMC",
                     "leafThickness", "SLA", "WD"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)

``` 


## 5. Perform calculations
```{r calculations }

gonzalez_processed_specieslvl <- cats_preparation(input_veg = gonzalez_veg, input_trait = gonzalez_trait, cutoff_value =  0.8, trait_information = trait_information,
                                  trait_detail = "Study")

mean_dist <- calculate_mean_dist(gonzalez_veg, gonzalez_processed_specieslvl)

FD <- dbFD(gonzalez_trait_fd, gonzalez_veg_fd, stand.x = T)

FD_df <- data.frame(FEve = FD$FEve,
        FDiv  = FD$FDiv,
        FRic = FD$FRic,
        RaoQ = FD$RaoQ,
        qual.FRic = FD$qual.FRic) %>%
  rownames_to_column(var = "Plot_code")

diversity_info <- gonzalez_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]] %>% 
  as.data.frame() %>% 
  mutate(shannon = diversity(.[,1:ncol(.)], index = "shannon", MARGIN = 1, base = exp(1)),
         simpson = diversity(.[,1:ncol(.)], index = "simpson", MARGIN = 1, base =exp(1)),
         species_richness = specnumber(.[,1:ncol(.)], MARGIN = 1),
         empty_states = ncol(gonzalez_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]]) - species_richness,
         meta_size = ncol(gonzalez_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]])) %>%
  select(shannon, simpson, species_richness, empty_states, meta_size) %>% # base = exp(1) gives the shannon index a log e base
  rownames_to_column(var = "Plot_code") %>% 
  full_join(mean_dist) %>%
  full_join(FD_df)


```



## 6. Cats model
```{r cats calculations}

#Determines amount of cores, substract four to keep to pc free for other processes.
n_cores <- parallel::detectCores() - 5 

my_cluster <- makeSOCKcluster(n_cores)

registerDoSNOW(cl = my_cluster)

foreach::getDoParRegistered() # check if cluster is set up


gonzalez_cats <- cats_calculations(gonzalez_processed_specieslvl, trait_information,
                                   trait_detail = "Study") %>%
  mutate(study = "Gonzalez")

gonzalez_figure <- cats_visualisation(gonzalez_cats) %>%
  mutate(study = "Gonzalez")

save(gonzalez_processed_specieslvl, gonzalez_figure, gonzalez_species_list, gonzalez_cats,
 file = "data/processed/gonzalez_calculations.RData")

ggplot(gonzalez_figure, aes(x = species_richness, y = KLR2)) +
  geom_point() +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm") + 
  labs(title = "The relationship between species richness and the CATS decomposition - Gonzalez et al. 20221",
       x = "Species richness") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.2, 1)) +
  scale_y_continuous(breaks = c(-0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  )) 

ggplot(gonzalez_figure, aes(x = mean_dist, y = KLR2)) +
  geom_point() +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm") + 
  labs(title = "The relationship between plot dissimilarity and the CATS decomposition - Gonzalez et al. 2021",
       x = "Average Bray-Curtis index value") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.2, 1)) +
  scale_y_continuous(breaks = c(-0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  )) 






```
