==> src/old_scripts/3. Complete_kassaw_V2.Rmd <==
---
title: "3. Complete analysis Kassaw et al."
author: "Jelyn Gerkema"
date: "2023-10-16"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 Kassaw et al. 2023
```{r clean data}

kassaw_veg <- read_csv("data/raw/Kassaw et al. 2023/species abundance.csv") %>%
  t() %>%
  as.data.frame() %>%
  rownames_to_column(var = "Name") %>%
  mutate(Name = gsub("_", " ", Name)) %>%
  mutate(Name = recode(Name, "Acacia abyssnica" = "Acacia abyssinica",
                       "Albizia gumifera" = "Albizia gummifera",
                       "Crateva adonsonia" = "Crateva adansonii",
                       "Olea europea subsp.cuspidata" = "Olea europaea subsp. cuspidata",
                       "Senna serguana" = "Senna singueana",
                       "Terminalia laxifolia" = "Terminalia laxiflora",
                       "Ziziphus spina" = "Ziziphus spina-christi")) %>%
  column_to_rownames(var = "Name") %>%
  t() %>%
  as.data.frame() %>%
  select(order(colnames(.)))

kassaw_trait <- read_csv("data/raw/Kassaw et al. 2023/trait data.csv") %>%
  rename(Name = species) %>%
  select(order(colnames(.))) %>%
  mutate(Name = gsub("_", " ", Name)) %>%
    mutate(Name = recode(Name, "Acacia abyssnica" = "Acacia abyssinica",
                       "Albizia gumifera" = "Albizia gummifera",
                       "Crateva adonsonia" = "Crateva adansonii",
                       "Olea europea subsp.cuspidata" = "Olea europaea subsp. cuspidata",
                       "Senna serguana" = "Senna singueana",
                       "Terminalia laxifolia" = "Terminalia laxiflora",
                       "Ziziphus spina" = "Ziziphus spina-christi")) %>%
  relocate(Name) %>%
  arrange(Name)


kassaw_trait_fd <- kassaw_trait %>%
  column_to_rownames(var = "Name") 

kassaw_trait <- kassaw_trait %>%
  mutate(across(where(is.numeric), ~ c(scale(., center = F)))) 

traitlist <- sort(c( "LA", "LDMC", "LT", "SLA"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)


kassaw_output <- lst(input_veg = kassaw_veg, 
                     input_veg_fd = kassaw_veg,
                     input_trait = kassaw_trait,
                     input_trait_fd = kassaw_trait_fd,
                     cutoff_value = 0.8,
                     trait_detail = "Study",
                     trait_information = trait_information,
                     study_name = "Kassaw")

save(kassaw_output,
 file = "data/processed/kassaw_prep.RData")

``` 


## 2. Perform calculations
```{r calculations }




########## later verwijderen
kassaw_specieslvl <- complete_analysis(input_veg = kassaw_veg, 
                            input_trait = kassaw_trait,
                            input_trait_fd = kassaw_trait_fd,
                            cutoff_value = 0.8,
                            trait_detail = "Study",
                            trait_information = trait_information,
                            study_name = "Kassaw"
                            )


test <- complete_analysis(input_veg = kassaw_output[["input_veg"]], 
                          input_veg_fd = kassaw_output[["input_veg_fd"]],
                          input_trait = kassaw_output[["input_trait"]],
                          input_trait_fd = kassaw_output[["input_trait_fd"]],
                          cutoff_value = kassaw_output[["cutoff_value"]],
                          trait_detail = kassaw_output[["trait_detail"]],
                          trait_information = kassaw_output[["trait_information"]],
                          study_name = kassaw_output[["study_name"]])




```
