==> src/old_scripts/3. Complete_marteinsdóttir.Rmd <==
---
title: "3. Complete analysis Eallonardo et al."
author: "Jelyn Gerkema"
date: "2023-06-20"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 Marteinsdottir et al. 2013
```{r clean data}

marteinsdottir_veg <- read_excel("data/raw/CESTES/Plant_data/Marteinsdottir2014_AJ.xlsx", 
    sheet = "commfull") %>%
    column_to_rownames(var = "Sites" )



marteinsdottir_info <- read_excel("data/raw/CESTES/Plant_data/Marteinsdottir2014_AJ.xlsx", 
    sheet = "envirfull") %>%
  mutate(across(Sites, as.character))

marteinsdottir_info2 <- read_excel("data/raw/CESTES/Plant_data/Marteinsdottir2014_AJ.xlsx", 
    sheet = "sitelist") %>%
  rename(Site_nr = Sites,
         Sites = SiteID) %>%
    mutate(across(Sites, as.character))

marteinsdottir_info <- marteinsdottir_info %>%
  full_join(marteinsdottir_info2) %>%
  rename(Plot_code = Sites) 

marteinsdottir_species_list <- read_excel("data/raw/CESTES/Plant_data/Marteinsdottir2014_AJ.xlsx", 
    sheet = "splist") %>%
  rename(Sp = TaxCode, Name = Taxon) %>%
  mutate(Name = gsub("\\.", " ", Name)) %>%
  mutate(Name = recode(Name, "Hieracium gr  Umbellatum" = "Hieracium umbellatum"))

colnames(marteinsdottir_veg) <- marteinsdottir_species_list$Name

marteinsdottir_veg <- marteinsdottir_veg %>%
  select(order(colnames(.)))
    
marteinsdottir_trait <- read_excel("data/raw/CESTES/Plant_data/Marteinsdottir2014_AJ.xlsx", 
  sheet = "traitsfull") %>%
  full_join(marteinsdottir_species_list) %>% 
  select(-c(SEED, COUNT, Sp)) %>%
  select(order(colnames(.))) %>%
  relocate(Name) %>%
  arrange(Name) 

marteinsdottir_trait_fd <- marteinsdottir_trait %>%
  column_to_rownames(var = "Name")

marteinsdottir_trait <- marteinsdottir_trait %>%
  mutate(across(where(is.numeric), ~ c(scale(., center = F)))) 


traitlist <- sort(c( "LDMC", "SLA"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)


``` 


## 5. Perform calculations
```{r calculations }

marteinsdottir_processed_specieslvl <- cats_preparation(input_veg = marteinsdottir_veg, 
                                                     input_trait = marteinsdottir_trait, cutoff_value =  0.8, 
                                                     trait_information = trait_information, 
                                                     trait_detail = "Study")

mean_dist <- calculate_mean_dist(marteinsdottir_veg, marteinsdottir_processed_specieslvl)

FD <- dbFD(marteinsdottir_trait_fd, marteinsdottir_veg,  stand.x = T)

FD_df <- data.frame(FEve = FD$FEve,
        FDiv  = FD$FDiv,
        FRic = FD$FRic,
        RaoQ = FD$RaoQ,
        qual.FRic = FD$qual.FRic) %>%
  rownames_to_column(var = "Plot_code")

diversity_info <- marteinsdottir_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]] %>% 
  as.data.frame() %>% 
  mutate(shannon = diversity(.[,1:ncol(.)], index = "shannon", MARGIN = 1, base = exp(1)),
         simpson = diversity(.[,1:ncol(.)], index = "simpson", MARGIN = 1, base =exp(1)),
         species_richness = specnumber(.[,1:ncol(.)], MARGIN = 1),
         empty_states = ncol(marteinsdottir_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]]) - species_richness,
         meta_size = ncol(marteinsdottir_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]])) %>%
  select(shannon, simpson, species_richness, empty_states, meta_size) %>% # base = exp(1) gives the shannon index a log e base
  rownames_to_column(var = "Plot_code") %>%
  full_join(mean_dist) %>%
  full_join(FD_df)


```



## 6. Cats model
```{r cats calculations}

marteinsdottir_cats <- cats_calculations(marteinsdottir_processed_specieslvl, trait_information,
                                         trait_detail = "Study") %>%
  mutate(study = "Marteinsdottir")

marteinsdottir_figure <- cats_visualisation(marteinsdottir_cats) %>%
  mutate(study = "Marteinsdottir")

save(marteinsdottir_processed_specieslvl, marteinsdottir_figure, marteinsdottir_species_list, 
     marteinsdottir_info, marteinsdottir_cats,
  file = "data/processed/marteinsdottir_calculations.RData")

marteinsdottir_figure <- marteinsdottir_figure %>%
  full_join(marteinsdottir_info)

ggplot(marteinsdottir_figure, aes(x = species_richness, y = KLR2)) +
  geom_point(aes(colour = Type)) +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm") + 
  labs(title = "The relationship between species richness and the CATS decomposition - Marteinsdóttir and Eriksson, 2013",
       x = "Species richness") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.2, 1)) +
  scale_y_continuous(breaks = c(-0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  )) 

ggplot(marteinsdottir_figure, aes(x = mean_dist, y = KLR2)) +
  geom_point(aes(colour = Type)) +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm") + 
  labs(title = "The relationship between plot dissimilarity and the CATS decomposition - Marteinsdóttir and Eriksson, 2013",
       x = "Average Bray-Curtis index value") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.2, 1)) +
  scale_y_continuous(breaks = c(-0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  )) 


```
