==> src/old_scripts/3. Complete_raevel.Rmd <==
---
title: "3. Complete analysis Raevel et al. 2012"
author: "Jelyn Gerkema"
date: "2023-06-20"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 Raevel et al. 2012
```{r clean data}

raevel_veg <- read_excel("data/raw/CESTES/Plant_data/Raevel2012_AJ.xlsx", 
    sheet = "comm") %>%
    column_to_rownames(var = "Sites") 

raevel_species_list <- read_excel("data/raw/CESTES/Plant_data/Raevel2012_AJ.xlsx", 
    sheet = "splist") %>%
  rename(Sp = TaxCode, Name = TaxName) %>%
   mutate(Name = paste0(tools::toTitleCase(word(Name, 1)), " ", word(Name, 2, -1))) %>%
  mutate(Name = recode(Name, "Dianthus longicaulis" = "Dianthus sylvestris subsp. longicaulis",
                       "Scabiosa atropurpurea var. maritima" = "Scabiosa atropurpurea" ))

colnames(raevel_veg) <- raevel_species_list$Name

raevel_veg <- raevel_veg %>%
  select(order(colnames(.)))

    
raevel_trait <- read_excel("data/raw/CESTES/Plant_data/Raevel2012_AJ.xlsx", 
  sheet = "traits") %>%
  full_join(raevel_species_list) %>% 
  rename(Height = H ) %>%
  select(-c(Sp, DM, LS, LF, FS,  CodeOri, Authorities, ScientificName, Family, Rank)) %>%
  select(order(colnames(.))) %>%
  relocate(Name) %>%
  arrange(Name) 

raevel_trait_fd <- raevel_trait %>%
  column_to_rownames(var = "Name")

raevel_trait <- raevel_trait %>%
  mutate(across(where(is.numeric), ~ c(scale(., center = FALSE)))) 




traitlist <- sort(c( "Height", "SLA", "SM"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)


``` 


## 5. Perform calculations
```{r calculations }

raevel_processed_specieslvl <- cats_preparation(input_veg = raevel_veg, input_trait = raevel_trait, cutoff_value =  0.8, trait_information = trait_information,
                                             trait_detail = "Study")


mean_dist <- calculate_mean_dist(raevel_veg, raevel_processed_specieslvl)

FD <- dbFD(raevel_trait_fd, raevel_veg,  stand.x = T)

FD_df <- data.frame(FEve = FD$FEve,
        FDiv  = FD$FDiv,
        FRic = FD$FRic,
        RaoQ = FD$RaoQ,
        qual.FRic = FD$qual.FRic) %>%
  rownames_to_column(var = "Plot_code")

diversity_info <- raevel_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]] %>% 
  as.data.frame() %>% 
  mutate(shannon = diversity(.[,1:ncol(.)], index = "shannon", MARGIN = 1, base = exp(1)),
         simpson = diversity(.[,1:ncol(.)], index = "simpson", MARGIN = 1, base =exp(1)),
         species_richness = specnumber(.[,1:ncol(.)], MARGIN = 1),
         empty_states = ncol(raevel_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]]) - species_richness,
         meta_size = ncol(raevel_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]])) %>%
  select(shannon, simpson, species_richness, empty_states, meta_size) %>% # base = exp(1) gives the shannon index a log e base
  rownames_to_column(var = "Plot_code") %>%
  full_join(mean_dist) %>%
  full_join(FD_df)



```



## 6. Cats model
```{r cats calculations}

raevel_cats <- cats_calculations(raevel_processed_specieslvl, trait_information,
                             trait_detail = "Study") %>%
  mutate(study = "Raevel")

raevel_figure <- cats_visualisation(raevel_cats) %>%
  mutate(study = "Raevel")

save(raevel_processed_specieslvl, raevel_figure, raevel_species_list, raevel_cats,
  file = "data/processed/raevel_calculations.RData")

ggplot(raevel_figure, aes(x = mean_dist, y = KLR2)) +
  geom_point() +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm") + 
  labs(title = "The relationship between plot dissimilarity and the CATS decomposition - Raevel et al., 2012",
       x = "Average Bray-Curtis value") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(0, 1)) +
  scale_y_continuous(breaks = c(-0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  )) 

ggplot(raevel_figure, aes(x = species_richness, y = KLR2)) +
  geom_point() +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm") + 
  labs(title = "The relationship between species richness and the CATS decomposition - Raevel et al., 2012",
       x = "Species richness") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(0, 1)) +
  scale_y_continuous(breaks = c(-0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  )) 


```
