==> src/old_scripts/4. Meta_analysis.Rmd <==
---
title: "Combine everything"
author: "Jelyn Gerkema"
date: "2023-10-20"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings.R")
source("src/functions.R")

```

## 1. Load data
```{r load data}

load("~/0. PhD/0. Chapter I/data/processed/aiello1992_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/aiello2011_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/bagaria_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/bergholz_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/buzzard_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/castro_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/chalmandrier_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/choler_calculations.RData")
#load("~/0. PhD/0. Chapter I/data/processed/cornwell_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/eallonardo_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/frenette2009_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/frenette2010_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/jaschke_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/kembel_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/liane_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/marteinsdottir_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/raevel_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/riva_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/rolhauser_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/kassaw_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/gonzalez_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/standen_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/makelele_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/jager_calculations.RData")
load("~/0. PhD/0. Chapter I/data/processed/osuri_calculations.RData")

meta_data <- read_excel("docs/data/Overview_datasets.xlsx")
load("~/0. PhD/0. Chapter I/data/processed/try_traits_tax_final.RData")
load("~/0. PhD/0. Chapter I/data/processed/try_save4112.RData")


```


```{r big compilation}

test <- bind_rows(aiello1992_figure, aiello2011_figure, bagaria_figure,
             buzzard_figure, castro_figure, chalmandrier_figure, choler_figure,
             eallonardo_figure, frenette2009_figure, frenette2010_figure,
             jaschke_figure, kembel_figure, liane_figure, marteinsdottir_figure,
             raevel_figure, riva_figure, rolhauser_figure, kassaw_figure, gonzalez_figure,
             standen_figure, makelele_figure) 

test <- test %>% full_join(meta_data)

test2 <- test %>%
   # filter(pval.meta < 0.051) %>%
  filter(study != "Li" & study != "Cornwell") 



ggplot(test2, aes(x = species_richness, y = KLR2, colour = study, shape = Habitat_short)) +
  geom_point(alpha = 0.3) +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm", se = F) + 
  labs(title = "The relationship between species richness and the CATS decomposition",  
       x = "Species richness") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.4, 1)) +
  scale_y_continuous(breaks = c(-0.4, -0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  )) + labs(col = "Dataset", shape = "Habitat type")

ggplot(test2, aes(x = mean_dist, y = KLR2, colour = study, shape = Habitat_short)) +
  geom_point(alpha = 0.3) +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm", se = F) + 
  labs(title = "The relationship between plot dissimilarity and the CATS decomposition",  
       x = "Average Bray-Curtis index value") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.4, 1)) +
  scale_y_continuous(breaks = c(-0.4, -0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  )) +
  scale_x_continuous(breaks = c(0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1)) +
  labs(col = "Dataset", shape = "Habitat type")


ggplot(test2, aes(x = log(FRic), y = KLR2, colour = study, shape = Habitat_short)) +
  geom_point(alpha = 0.3) +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm", se = F) + 
  labs(title = "The relationship between functional richness and the CATS decomposition",  
       x = "Functional richness (log)") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.4, 1)) +
  scale_y_continuous(breaks = c(-0.4, -0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  )) +
  labs(col = "Dataset", shape = "Habitat type")

ggplot(test2, aes(x = FDiv, y = KLR2, colour = study, shape = Habitat_short)) +
  geom_point(alpha = 0.3) +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm", se = F) + 
  labs(title = "The relationship between functional richness and the CATS decomposition",  
       x = "Functional richness (log)") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.4, 1)) +
  scale_y_continuous(breaks = c(-0.4, -0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  )) +
  labs(col = "Dataset", shape = "Habitat type")

ggplot(test2, aes(x = FEve, y = KLR2, colour = study, shape = Habitat_short)) +
  geom_point(alpha = 0.3) +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm", se = F) + 
  labs(title = "The relationship between functional richness and the CATS decomposition",  
       x = "Functional richness (log)") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.4, 1)) +
  scale_y_continuous(breaks = c(-0.4, -0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  )) +
  labs(col = "Dataset", shape = "Habitat type")


ggplot(test, aes(x = (species_richness), y = FRic)) +
  geom_point(alpha = 0.12, aes(colour = study)) +
  theme_classic() +
  geom_smooth(method = "lm", se = T)

```



```{r species list}

study_input <- sort(unique(test2$study))

study_lwc <- tolower(study_input)

study_lwc <- c("aiello1992", "aiello2011", "bagaria",  "buzzard", "castro", "chalmandrier", "choler", "eallonardo", "frenette2009", "frenette2010", "gonzalez", "jager", "jäschke", "kassaw", "kembel", "liane", "makelele", "marteinsdottir", "osuri", "raevel", "riva", "rolhauser", "standen"
)


data_input_list <- lst(aiello1992_processed_specieslvl, aiello2011_processed_specieslvl, bagaria_processed_specieslvl,  buzzard_processed_specieslvl, castro_processed_specieslvl, chalmandrier_processed_specieslvl, choler_processed_specieslvl, eallonardo_processed_specieslvl, frenette2009_processed_specieslvl, frenette2010_processed_specieslvl, gonzalez_processed_specieslvl,
jager_processed_specieslvl, jaschke_processed_specieslvl, kassaw_processed_specieslvl, kembel_processed_specieslvl, liane_processed_specieslvl, makelele_processed_specieslvl, marteinsdottir_processed_specieslvl, osuri_processed_specieslvl, raevel_processed_specieslvl, riva_processed_specieslvl, rolhauser_processed_specieslvl, standen_processed_specieslvl)

output <- data.frame()

for(i in 1:length(data_input_list)){
  
  input <- data_input_list[[i]]
  
  n_plot_before <- ncol(input[["relative_abundances"]][["matrix"]])
  
  n_plot_after <- ncol(input[["relative_abundances"]][["matrix4traits_prop"]])
  
  study <- study_lwc[[i]]
  
  output1 <- bind_cols(n_plot_before = n_plot_before, n_plot_after =  n_plot_after, study = study)
  
  output <- bind_rows(output, output1)
  
}


for(i in 1:length(data_input_list)){
  
  
  input <- data_input_list[[i]]
  
  species <- colnames(input[["relative_abundances"]][["matrix4traits_prop"]]) %>% 
  as.data.frame() %>%
  rename(Name = ".") %>%
  mutate(study = study_lwc[i])
  
  output <- bind_rows(output, species)
  
  
}

Oké ik moet even nadenken over de workflow. Ik heb de taxonomische informatie van alle soorten nodig
per 


output_tmp2 <- output %>% 
  #select(-c(study)) %>%
  distinct() %>%
  arrange(Name) %>%
  #pull(Name) %>%
  rename(dataset_name = Name)

total_species_list <- output_tmp2 %>%
  select(dataset_name) %>%
  distinct()

try_tax_final_clean <- try_traits_tax_final %>%
  rename(old_try_name = AccSpeciesName,
         new_wfo_name = new_name)

#These are the species that are present within the new_name list n = 1038
species_present <- total_species_list %>%
  filter(dataset_name %in% c(try_tax_final_clean$new_wfo_name)) %>%
  mutate(new_wfo_name = dataset_name) %>%
  merge(try_tax_final_clean) %>%
  select(new_wfo_name, dataset_name, old_try_name, acceptedNameUsageID) %>%
  distinct()

length(unique(species_present$dataset_name))


# These are the species that are not present in the new_name list
# but do have an entry within the original try species names
species_present2 <- total_species_list %>%
  filter(dataset_name %notin% c(species_present$dataset_name)) %>%
  filter(dataset_name %in% c(try_tax_final_clean$old_try_name)) %>%
  mutate(old_try_name = dataset_name) %>%
  merge(try_tax_final_clean) %>%
  select(new_wfo_name, dataset_name, old_try_name, acceptedNameUsageID) %>%
  distinct()


# In total, 1062 species are present in the filtered try dataset
species_present_both <- bind_rows(species_present, species_present2)

length(unique(species_present_both$dataset_name))

# That means that there are 439 species that are not present, maybe they are present
# under a synonym
not_present <- total_species_list %>%
  filter(dataset_name %notin% c(species_present_both$dataset_name))
  
# Of the species that are not present, 87 have a synonym in the wfo database 
# of species names and in the cleaned try database
not_present_id <- classification %>%
  mutate(scientificName = str_replace_all(scientificName, "  ", " ")) %>%
  filter(scientificName %in% c(not_present$dataset_name)) %>%
  drop_na(acceptedNameUsageID) %>%
  select(acceptedNameUsageID, scientificName) %>%
  rename(dataset_name = scientificName) %>%
  distinct() %>%
  filter(acceptedNameUsageID %in% c(try_tax_final_clean$acceptedNameUsageID)) %>%
  group_by(dataset_name) %>%
  filter(n() == 1) %>%
  ungroup() 

length(unique(id_present$acceptedNameUsageID))

id_present <- not_present_id %>%
  merge(try_tax_final_clean) %>%
  select(new_wfo_name, dataset_name, old_try_name, acceptedNameUsageID) %>%
  distinct()

length(unique(combined_present$dataset_name))

## In total, 1110 species are present in the try database. 
combined_present <- bind_rows(species_present_both, id_present)

combined_present_full <- try_tax_final_clean %>%
  select(genus, family, new_wfo_name, acceptedNameUsageID) %>%
  distinct() %>%
  right_join(combined_present) 

combined_present_tax <- combined_present_full %>%
  select(dataset_name, genus, family) %>%
  distinct()


classification_temp <- classification %>%
   mutate(scientificName = str_replace_all(scientificName, "  ", " ")) %>%
  rename(dataset_name = scientificName)

not_included_final <- total_species_list %>%
  filter(dataset_name %notin% c(combined_present_full$dataset_name)) %>%
  left_join(classification_temp) %>%
  select(genus, family, dataset_name) %>%
  distinct()


final_final <- bind_rows(combined_present_tax, not_included_final) %>%
  left_join(output_tmp2) 



no_data <- final_final %>%
  filter(is.na(genus)) %>%
  separate(dataset_name, into = c("genus"), sep = " ", extra = "drop", remove = F) %>%
  select(-family) 

family_data <- classification %>%
  filter(genus %in% c(no_data$genus)) %>%
  filter(taxonomicStatus != "Synonym") %>%
  select(genus, family) %>%
  distinct() 


no_data_filled <- no_data %>%
  full_join(family_data) %>%
  mutate(genus = case_when(dataset_name %in% c("Asteraceae sp. 2",
                                               "Euphorbiaceae sp. 1",
                                               "Rosaceae sp. 1",
                                               "Rubiaceae sp. 4",
                                               "Unidentified species 6") ~ NA,
                           TRUE ~ genus)) %>%
  mutate(family = case_when(dataset_name == "Asteraceae sp. 2" ~ "Asteraceae",
                            dataset_name == "Euphorbiaceae sp. 1" ~ "Euphorbiaceae",
                            dataset_name == "Rosaceae sp. 1"  ~ "Rosaceae",
                            dataset_name == "Rubiaceae sp. 4" ~ "Rubiaceae",
                            genus == "Buddleia" ~ "Scrophulariaceae",
                            genus == "Drymis" ~ "Winteraceae",
                            TRUE ~ family))
 

final_final_final <- final_final %>%
  filter(dataset_name %notin% c(no_data_filled$dataset_name)) %>%
  bind_rows(no_data_filled)

genus_count <- final_final_final %>%
  select(c(study, genus))  %>%
  distinct()

family_count <- final_final_final %>%
  select(c(study, family))  %>%
  distinct()



genus_count <-  table(genus_count$study) %>% as.data.frame() %>%
  rename(study = Var1, genus_n = Freq)
species_count <- table(final_final_final$study) %>% as.data.frame() %>%
  rename(study = Var1, species_n = Freq)
family_count <- table(family_count$study) %>% as.data.frame() %>%
  rename(study = Var1, family_n = Freq)

taxonomic_counts <- lst(genus_count, species_count, family_count) %>%
  reduce(full_join)

coverage_output <- data.frame()

for(i in 1:length(data_input_list)){
  
  input <- data_input_list[[i]]
  
  regional_trait_coverage_try <- input[["relative_abundances"]][["matrix"]] %>% 
    as.data.frame() %>%
    summarise(across(where(is.numeric), ~ sum(.))) %>%
    t() %>%
    as.data.frame() %>%
    mutate(Total_cover = rowSums(.)) %>%
    select(Total_cover) %>%
    filter(Total_cover != 0) %>%
    mutate(Relative_meta_cover_try = Total_cover / sum(Total_cover)) %>%
    rownames_to_column(var = "Name") %>%
    filter(Name %in% combined_present_tax$dataset_name) %>%
    select(Name, Relative_meta_cover_try) %>%
    summarise(across(where(is.numeric), ~sum(.))) %>%
    mutate(study = study_lwc[i])
  
  regional_trait_coverage <- input[["relative_abundances"]][["matrix"]] %>% 
    as.data.frame() %>%
    summarise(across(where(is.numeric), ~ sum(.))) %>%
    t() %>%
    as.data.frame() %>%
    mutate(Total_cover = rowSums(.)) %>%
    select(Total_cover) %>%
    filter(Total_cover != 0) %>%
    mutate(Relative_meta_cover_orig = Total_cover / sum(Total_cover)) %>%
    rownames_to_column(var = "Name") %>%
    filter(Name %in% colnames(input[["relative_abundances"]][["matrix_prop_match"]])) %>%
    select(Name, Relative_meta_cover_orig) %>%
    summarise(across(where(is.numeric), ~sum(.))) %>%
    mutate(study = study_lwc[i]) %>%
    full_join(regional_trait_coverage_try)
  
  coverage_output <- bind_rows(coverage_output, regional_trait_coverage)
  
  
}

genus_prep <-   kassaw_processed_specieslvl[["relative_abundances"]][["matrix"]] %>% 
    as.data.frame() %>% 
    rownames_to_column(var = "Plot_code") %>%
    pivot_longer(cols = 2:ncol(.), values_to = "abund", names_to = "dataset_name") 

genus_lvl_test <- final_final_final %>%
  filter(study == "kassaw") %>%
  full_join(genus_prep) %>%
  select(-c(dataset_name, family, study)) %>%
  group_by(genus, Plot_code) %>%
  summarize(across(where(is.numeric), ~sum(abund))) %>%
  ungroup() %>%
  pivot_wider(values_from = "abund", names_from = "genus") %>%
  column_to_rownames(var = "Plot_code")



coverage_output_tmp <- coverage_output %>%
  relocate(Relative_meta_cover_orig, Relative_meta_cover_try, study) %>%
  mutate(check = case_when(Relative_meta_cover_try >= 0.8 ~ "Yes",
                           Relative_meta_cover_try <0.8 ~ "No"))





# 452 are not present at all in the cleaned try database. 
really_not_present <- total_species_list %>%
  filter(dataset_name %notin% c(combined_present$dataset_name)) 








``` 

### impact of bias
```{r cats bias}

colnames(aiello1992_cats)

cats_list <- lst(aiello1992_cats, aiello2011_cats, bagaria_cats, bergholz_cats, buzzard_cats, castro_cats, chalmandrier_cats, choler_cats, eallonardo_cats, frenette2009_cats, frenette2010_cats, gonzalez_cats, jaschke_cats, kassaw_cats, kembel_cats, liane_cats, makelele_cats, marteinsdottir_cats, raevel_cats, riva_cats, rolhauser_cats, standen_cats)

figure_list <- lst(aiello1992_figure, aiello2011_figure, bagaria_figure, bergholz_figure, buzzard_figure, castro_figure, chalmandrier_figure, choler_figure, eallonardo_figure, frenette2009_figure, frenette2010_figure, gonzalez_figure, jaschke_figure, kassaw_figure, kembel_figure, liane_figure, makelele_figure, marteinsdottir_figure, raevel_figure, riva_figure, rolhauser_figure, standen_figure)

output <- data.frame()

for(i in 1:length(cats_list)){
  
  
  input_cats <- cats_list[[i]]
  input_figure <- figure_list[[i]]
  
  cats <- input_cats %>% 
  select(`model.bias`, `model.uniform.prior.plus.traits`, `model.mean.null.given.uniform`,
         `model.mean.null.given.prior`, `model.prior.plus.traits`, `Plot_code`, `study`)
  
  diversity <- input_figure %>%
  select(species_richness, shannon, simpson, empty_states, meta_size, mean_dist, study, Plot_code) %>%
  distinct() %>%
  full_join(cats)
  
  output <- bind_rows(output, diversity)
  
  
}

output2 <- output %>%
  pivot_longer(cols = 9:12, values_to = "Value", names_to = "Model")

output3 <- output %>%
  mutate(just.neutral1 = model.prior.plus.traits - model.uniform.prior.plus.traits,
      just.neutral2 = model.mean.null.given.prior - model.mean.null.given.uniform,
      just.traits1 = model.uniform.prior.plus.traits - model.mean.null.given.uniform,
      just.traits2 = model.prior.plus.traits - model.mean.null.given.prior,
      information.unique.to.local.trait.constraints = just.traits2 ,
      information.unique.to.neutral.prior = just.neutral1 ,
      delta.R.trait = just.traits1,
      delta.R.neutral = just.neutral2,
      T1 = just.traits1 - just.traits2,
      T2 = just.neutral2 - just.neutral1,
      joint.information = T2 ,
      biologically.unexplained.information = (1 - model.prior.plus.traits)) %>%
  select(-c(T1, T2, just.neutral1, just.neutral2, just.traits1, just.traits2,
            delta.R.trait, delta.R.neutral)) %>%
  pivot_longer(cols = 14:17, values_to = "KLR2", names_to = "Deviance") %>%
  mutate(Deviance = recode(Deviance,
                             biologically.unexplained.information = "Unexplained effect",
                             information.unique.to.neutral.prior = "Pure metacommunity effect",
                             joint.information = "Joint trait & metacommunity effect",
                             information.unique.to.local.trait.constraints = "Pure trait effect"
    ))



ggplot(output3, aes(x =  `species_richness`, y = KLR2, colour = study)) +
  geom_point(alpha = 0.3) +
  theme_classic() +
  geom_smooth(method = "lm", se = F) + 
  labs(title = "CATS decomposition, uncorrected for bias",
       x = "Species richness") +
  labs(col = "Dataset") + facet_wrap(~Deviance) +
    geom_hline(yintercept = 0, linetype = 2) 

output4 <- full_join(output, meta_data)

ggplot(output4, aes(x = `species_richness`, y = `model.bias`, colour = study,
                   shape = Habitat_short)) + 
  geom_point(alpha = 0.3) +
  theme_classic() +
  geom_smooth(method = "lm", se = F) +
  labs(title = "Relationship between model bias and species richness (distribution length)",
       x = "Species richness", y = "Model bias", col = "Dataset", shape = "Habitat type") 

ggplot(output2, aes(x =  `empty_states`, y = Value, colour = study)) +
  geom_point(alpha = 0.3) +
  theme_classic() +
  geom_smooth(method = "lm", se = F) + 
  labs(title = "  ") +
  labs(col = "Dataset") + facet_wrap(~Model)
```



```{r figures}
assigned_colnames <- paste0("species_", letters[1:20])

community_a <- c(rep(0.125, 8), rep(NA, 20-8)) %>% t() %>% as.data.frame()
colnames(community_a) <- assigned_colnames

community_b <- c(rep((1/10), 10), rep(NA, 20-10)) %>% t() %>% as.data.frame()
colnames(community_b) <- assigned_colnames

community_c <- c(rep((1/4), 4), rep(NA, 20-4)) %>% t() %>% as.data.frame()
colnames(community_c) <- assigned_colnames

community_d <- c(rep((1/12), 12), rep(NA, 20-12)) %>% t() %>% as.data.frame()
colnames(community_d) <- assigned_colnames

community_e <- rep((1/20), 20) %>% t() %>% as.data.frame()
colnames(community_e) <- assigned_colnames


community_df <- bind_rows(community_c, community_a, community_b,
                          community_d, community_e) %>%
  rownames_to_column(var = "Community") %>%
  pivot_longer(cols = 2:ncol(.), values_to = "Probability", names_to = "Species") %>%
  mutate(Species = as.character(Species) %>% fct_reorder(Species))


ggplot(community_df, aes(x = Species, y = Probability, colour = Community, group = Community)) +
  geom_point() + geom_line() + theme_classic() +
  theme(axis.text.x = element_text(angle = 70, vjust = 0.5)) +
  labs(title = "Uniform species probability distributions (1/S)") +
   coord_cartesian(ylim = c(0, 0.3))


random_numbers <- runif(20)


probability_distribution <- random_numbers / sum(random_numbers)

community_f <- probability_distribution %>% t() %>% as.data.frame()

community_g <-  probability_distribution %>% t() %>% as.data.frame()

community_h <-  c(0.2, 0.15, 0.1, 0.07, 0.06, 0.03, rep((0.39/14), 14))  %>% t() %>% 
as.data.frame()
colnames(community_f) <- assigned_colnames
colnames(community_g) <- assigned_colnames
colnames(community_h) <- assigned_colnames


random_numbers <- abs(rnorm(20))

probability_distribution <- random_numbers / sum(random_numbers)

community_i <-  probability_distribution %>% t() %>% as.data.frame()
community_j <-  c(rep((0.23/12), 12),
                 0.02, 0.02, 0.03, 0.03, 
                 0.07, 0.1, 0.2, 0.3) %>% t() %>% as.data.frame()

colnames(community_i) <- assigned_colnames
colnames(community_j) <- assigned_colnames




community_random <- bind_rows(community_i) %>%
  rownames_to_column(var = "Community") %>%
  pivot_longer(cols = 2:ncol(.), values_to = "Probability", names_to = "Species") %>%
  mutate(Species = as.character(Species) %>% fct_reorder(Species))


ggplot(community_random, aes(x = Species, y = Probability, colour = Community, group = Community)) +
  geom_point() + geom_line() + theme_classic() +
  theme(axis.text.x = element_text(angle = 70, vjust = 0.5)) +
  labs(title = "Random species probability distributions", x = NULL) +
   coord_cartesian(ylim = c(0, 0.3))


community_o <-  c(0.3, 0.2, 0.1, 0.07, 0.06, 0.02, 0.02, rep((0.23/13), 13)) %>% 
  t() %>% as.data.frame()

community_p <- c(0.2, 0.1, 0.1, 0.03, 0.03, 0.1, 0.1, rep((0.66/13), 13)) %>% 
  t() %>% as.data.frame()

community_r <- c(0.34, 0.28, 0.11, 0.05, 0.05, 0.03, rep((0.14/13), 14)) %>% 
  t() %>% as.data.frame()

colnames(community_o) <- assigned_colnames
colnames(community_p) <- assigned_colnames
colnames(community_r) <- assigned_colnames

community_exam <- bind_rows(community_o) %>%
  rownames_to_column(var = "Community") %>%
  pivot_longer(cols = 2:ncol(.), values_to = "Probability", names_to = "Species") %>%
  mutate(Species = as.character(Species) %>% fct_reorder(Species))


ggplot(community_exam, aes(x = Species, y = Probability, colour = Community, group = Community)) +
  geom_point() + geom_line() + theme_classic() +
  theme(axis.text.x = element_text(angle = 70, vjust = 0.5)) +
  labs(title = "Random species probability distributions", x = NULL, colour = NULL) +
   coord_cartesian(ylim = c(0, 0.3))


```
